### 2019-05-09 #1436 @whedon
[PRE REVIEW]: mTower: Trusted Execution Environment for MCU-based devices
**Submitting author:** @omoliavko (<a href="http://orcid.org/0000-0003-3167-983X">Oleksandr Moliavko</a>)
**Repository:** <a href="https://github.com/Samsung/mTower.git" target ="_blank">https://github.com/Samsung/mTower.git</a>
**Version:** v0.1
