### 2019-05-09 #1436 @labarba
@omoliavko — Thank you for your submission to JOSS. From a quick inspection of the paper, it's not clear that it meets our [submission criteria](http://joss.theoj.org/about#submission_requirements). In particular:

> - Your software should have an obvious research application

You have a section header on "Research application" but the content is unconvincing—bear in mind we ask that the paper be written for a technical non-specialist. As such, it's unclear how security assessment of IoT devices is a research question. Can you clarify?
