### 2019-05-10 fca233e
Updated research applications part of JOSS paper draft
