+++ docs/joss/paper.md
@@ -28,6 +28,7 @@ Embedded computing systems have already become omnipresent in day-to-day life, a
 
 # Implementation overview
 In general, secure applications using TrustZone hardware as hardware security basis consist of two interacting components: Normal World (NW) and Secure World (SW). Normal World part is, in general, a standard Linux application that uses TEE library containing API functions to interact with Secure World part. Corresponding Secure World part of application is essentially a set of event handlers that are executed in hardware-protected area of RAM under control of specially designed operating system. Secure World part processes calls received from Normal World and deals with sensitive data that must be protected from unauthorized access (cryptographic keys, passwords, user identification data, etc.). Common operations performed by Secure World part of application include data encryption/decryption, user authentication, key generation, digital signing and signature verification. Overall architecture that applications should adhere to is described in TEE System Architecture v1.2 document [@TEESystemArch].
+
 mTower implementation follows this layout: it provides implementation of APIs described in GP TEE specification documents. Two most relevant TEE specifications are
 
 * TEE Client API Specification v1.0[@ClientAPISpec] describing the communication between NW applications and trusted applications residing in the SW.
@@ -50,7 +51,7 @@ mTower project is expected to produce proof-of-concepts and evaluation models fo
 * Computer security specialists
 
 # Research applications
-Development of mTower was initiated as a proof-of-concept project to demonstrate feasibility of porting GP TEE-based standard compliant trusted applications from full-feature CPU-based to minimalistic MCU-based systems. So far, only the simplest applications were ported and run on ARM Cortex-M23 family MCU; to run more complex applications, further development is needed (first and foremost, a support for shared memory buffer in API). Other potential research application for mTower is using it as platform for security assessment in Internet-of-things and EDGE devices. As EDGE initiative encompasses a wide spectrum of devices (industrial, healthcare, communication, automotive, etc.), mTower can potentially find a lot of research and evaluation applications.
+mTower is being developed as a part of search for efficient hardware/software security solution for embedded systems. It is expected to demonstrate feasibility of porting GP TEE-based standard compliant trusted applications from full-feature CPU-based to MCU-based systems. Other intended research application for mTower is using it as platform for developing and evaluating secure applications and frameworks suitable for resource-limited environments, including Internet-of-things and EDGE devices. So far, security issues were rarely addressed during development of applications for embedded systems, and there very little info on performance and resource consumption penalties incurred by adding security mechanisms into applications for such systems. We plan to measure these penalties (in the simplest case, by comparing results obtained from running profiling tool over two versions of same application, first with security mechanism based on mTower and second without security mechanism), to provide concrete numbers and, hopefully, prove that applying ARM TrustZone-based security solutions to embedded system applications is technically feasible. It should be noted that mTower will provide an assortment of cryptography- and security-related functions (user authentication, cryptographic key generation, encryption/decryption, etc.), and performance of each of these functions needs to be evaluated to assure their usability on embedded systems, as the applications on such systems are typically stripped down to use only necessary code, with as little overhead as possible.
 
 # References
 
