### 2019-05-10 #1436 @omoliavko
@labarba - thank you for comment. I've rewritten "Research application" part of article. Could you please take a look at it once more? If specific measurement results are needed, I suppose these can be obtained, but conditions under which the measurements will be performed will be artificial in any case.
