### 2019-05-10 #1436 @labarba
@omoliavko — Thanks for the update on your paper. The editors have discussed this submission on our email list, and remain hesitant about whether it is in-scope for JOSS. We say in our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) that

> The software should have an obvious research application

The goal is to publish articles that allow **citation** of these software works, when used by other researchers. Although you explain _your_ plans to use the software in some research, can you articulate how it may be used by _others_ in pursuit of answering a research question?
