### 2019-05-11 4b5a90c
Rewritten research applications part of JOSS paper draft
