+++ docs/joss/paper.md
@@ -51,7 +51,7 @@ mTower project is expected to produce proof-of-concepts and evaluation models fo
 * Computer security specialists
 
 # Research applications
-mTower is being developed as a part of search for efficient hardware/software security solution for embedded systems. It is expected to demonstrate feasibility of porting GP TEE-based standard compliant trusted applications from full-feature CPU-based to MCU-based systems. Other intended research application for mTower is using it as platform for developing and evaluating secure applications and frameworks suitable for resource-limited environments, including Internet-of-things and EDGE devices. So far, security issues were rarely addressed during development of applications for embedded systems, and there very little info on performance and resource consumption penalties incurred by adding security mechanisms into applications for such systems. We plan to measure these penalties (in the simplest case, by comparing results obtained from running profiling tool over two versions of same application, first with security mechanism based on mTower and second without security mechanism), to provide concrete numbers and, hopefully, prove that applying ARM TrustZone-based security solutions to embedded system applications is technically feasible. It should be noted that mTower will provide an assortment of cryptography- and security-related functions (user authentication, cryptographic key generation, encryption/decryption, etc.), and performance of each of these functions needs to be evaluated to assure their usability on embedded systems, as the applications on such systems are typically stripped down to use only necessary code, with as little overhead as possible.
+mTower is being developed as a part of search for efficient hardware/software security solution for embedded systems. It is expected to demonstrate feasibility of porting GP TEE-based standard compliant trusted applications from full-feature CPU-based to MCU-based systems. Other intended research application for mTower is using it as platform for developing and evaluating secure applications and frameworks suitable for resource-limited environments, including Internet-of-things and EDGE devices. So far, security issues were rarely addressed during development of applications for embedded systems, and there very little info on performance and resource consumption penalties incurred by adding security mechanisms into applications for such systems. The prevailing opinion seems to be that security through obscurity is sufficient for embedded systems, and any attempt to introduce sophisticated security features will incur unacceptable overhead. So one of the objectives of mTower development is to provide a reference platform which would allow the researchers to actually measure these penalties (in the simplest case, by comparing results obtained from running profiling tool over two versions of same application, first with security mechanism based on mTower and second without security mechanism). We hope that when specific, measurable performance data will become available, ARM TrustZone-based security solutions for embedded system applications will be seen as technically feasible. While we cannot foresee all potential use cases for mTower, and, therefore, cannot provide performance estimates for them, we plan to equip mTower with software hooks and built-in measurement mechanisms, so the potential users can easily measure performance of cryptographic and data security related mechanisms in any application they will base on mTower.
 
 # References
 
