### 2019-05-11 #1436 @omoliavko
@labarba - I apologize for not explaining potential research applications for mTower in more detail. I've updated the relevant part of paper once more to emphasize that mTower can be used by any developer interested in evaluating performance of TrustZone-based security mechanisms in embedded systems. If necessary, I can provide examples of potential usage scenarios of mTower (descriptions of test cases and  and measurement procedures), but these would be our own tests, not ones developed by third-party researchers.
