### 2019-05-20 8904543
Partially implemented (ported) the GP "TEE Internal API"
