+++ tee_client/public/tee_client_api.h
@@ -349,7 +349,8 @@ typedef struct {
  */
 typedef struct {
 	uint32_t a;
-	uint32_t b;
+  uint32_t b;
+  uint32_t c;
 } TEEC_Value;
 
 /**
