### 2019-05-30 c2aa234
Refactored code structure
