### 2019-05-30 4dcf392
Ported hotp example from OP-TEE
