### 2019-06-08 #1436 @gkthiruvathukal
@gradvohl, Would you be willing to contribute a review for this JOSS submission?
