### 2019-06-08 #1436 @gkthiruvathukal
@federeghe, Would you be willing to contribute a review for this JOSS submission?
