### 2019-06-08 #1436 @gradvohl
>
>
> @gradvohl, Would you be willing to contribute a review for this JOSS submission?

Yes. I am happy to review it.
