### 2019-06-09 #1494 @gradvohl
# Review items for JOSS paper #1494
**Paper title**: mTower: Trusted Execution Environment for MCU-based devices

## Review criteria
* A list of the authors of the software: :heavy_check_mark:
* Author affiliations: :heavy_check_mark:
* A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience: :heavy_check_mark:
* A clear statement of need that illustrates the purpose of the software: :heavy_check_mark:
* Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it: :heavy_check_mark:
* A list of key references including a link to the software archive: :heavy_exclamation_mark: :arrow_heading_up: (Needs improvements)

## Review itens
* Software license: :exclamation::question:

* Documentation:
    * A statement of need: :thumbsup: (good)
    * Installation instructions: :wave: (sort of)
    * Example usage: :thumbsdown: (bad)
    * API documentation: :thumbsup: (good)
    * Community guidelines: :thumbsup: (good)
    * Functionality: :confused:
    * Tests: :thumbsdown: (bad)

* Software paper
    * Authors: :exclamation::question:
    * A statement of need: :heavy_check_mark:
    * References: :heavy_exclamation_mark: :arrow_heading_up: (Needs improvements)

## Reviewer comments
The following comments refer to items that I considered were not met in full when I reviewed this submission.

The authors mention two licenses in the repository. While the repository itself uses the Apache 2.0 license, the README.md text, in section 2, indicates the BSD-2 license.

The first author made many contributions to the repository, although he was not the one who made the most contributions. Regarding the list of authors of the paper, it seems that one of them -- Oleg Kopysov -- did not contribute to the repository.

Regarding the references, I found only references to the specification of the TEE API or the TEE architecture. Still, to download these references, you need to fill out a form. I believe that the authors could add other references. For instance, references that better explore the security issues that mTower intends to address or references to discuss the importance of FreeRTOS as the operating system for these MCU-based devices.

Considering the examples of use, I have not found any well-documented examples that illustrate the operation of mTower. I saw that there are some directories with examples, but they are scattered in the repository and there is not a step-by-step guide to running the examples.

Concerning the functionalities, unfortunately, I do not have the specific platform the software runs to execute some tests. In this sense, it would be interesting -- although I do not consider it mandatory -- if the authors indicated a container or a virtual machine where the users could run some experiments.

Finally, I could not find mechanisms for automatic testing or test results. I suggest trying to add some tests and results reports.

## Conclusion
Analyzing all the aspects mentioned earlier, I classify the paper in the category of major revisions.
