### 2019-06-09 #1494 @federeghe
I checked the LICENSE and COPYING file verifying that all the used software is compliant with respect to the Apache license. However, not only the README file uses the BSD license (instead of the Apache of the LICENSE file) but also some source code files have BSD license in the header. Please make the license uniform across files and documentation.

I did not perform a in-depth review because I have the same concerns of @gradvohl. In particular:
- The submitter to JOSS is not the author that made a major contribution to the software. Actually, @omoliavko pushed commits only about the JOSS paper. However, I see that there are some large commits from @tdrozdovsky, that are possibly multi-author works. I think that a confirmation from @tdrozdovsky about the software authorship may be sufficient to solve this issue.
- There are some examples in `apps` folder, but it is not clear how to build them.
- Please expand the documentation on how to add a new architecture, in particular about the `apps` folder and related files.
- As outlined by @gradvohl, neither automated nor manual tests are present. They are required by the journal policy.
