### 2019-06-10 Samsung/mTower#4 @omoliavko
Fix license discrepancies
**Describe the bug**
There are several conflicting license types used in mTower components.

**To Reproduce**

**Expected behavior**
Licenses should be non-conflicting

**Additional context**
Apache license should be applied to mTower overall; however, some files will remain under BSD license maintaining the rights of third parties.
