### 2019-06-10 Samsung/mTower#5 @omoliavko
Fix author list in article
**Describe the bug**
Author list in article should reflect actual contribution order in project
