### 2019-06-10 Samsung/mTower#6 @omoliavko
Improve installation instructions
**Describe the bug**
mTower needs concise installation instructions.
