### 2019-06-10 Samsung/mTower#7 @omoliavko
A guide on creating mTower-based app is needed
**Describe the bug**
Need to create a specific, easy to follow guide on creating mTower-based secure application.

**Expected behavior**
A document must be present containing the said guide.
