### 2019-06-10 Samsung/mTower#8 @omoliavko
Add TEE documentation into repository and improve reference list
**Describe the bug**
Documentation on TEE and is difficult to access in online sources. Also, reference list in article is to be improved.

**Expected behavior**
Will move as much of TEE documentation as possible directly into repository; will improve reference list.
