### 2019-06-10 Samsung/mTower#9 @omoliavko
A guide for porting mTower to new H/W platforms is absent
**Describe the bug**
mTower needs a guide on porting it to other H/W platforms.
