### 2019-06-10 Samsung/mTower#10 @omoliavko
GP TEE API test suite for mTower is needed
**Describe the bug**
There's no test suite to demonstrate that GP TEE API implementation in mTower operates as it should.

**Expected behavior**
A test suite must illustrate operation of standard GP TEE API calls and their usage.

**mTower**
* Platform: nuvoton NuMaker-PFM 2351,
* version: v0.1
* Toolchain: GNU GCC
