### 2019-06-21 78f7a5f
[Bug #4] Fixed license discrepancies
