### 2019-06-21 4345751
[Bug #4] Fixed license discrepancies
