+++ tee/include/kernel/pseudo_ta.h
@@ -1,3 +1,22 @@
+/**
+ * @file        tee/include/kernel/pseudo_ta.h
+ *
+ * @copyright   Copyright (c) 2019 Samsung Electronics Co., Ltd. All Rights Reserved.
+ * @author      Taras Drozdovskyi t.drozdovsky@samsung.com
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 /* SPDX-License-Identifier: BSD-2-Clause */
 /*
  * Copyright (c) 2015, Linaro Limited
@@ -40,7 +59,6 @@ struct pseudo_ta_head {
 
 #define pseudo_ta_register(...) static const struct pseudo_ta_head __head \
 			__used __section("ta_head_section") = { __VA_ARGS__ }
-
 
 struct pseudo_ta_ctx {
 	const struct pseudo_ta_head *pseudo_ta;
