### 2019-06-21 77d81c0
[Bug #5] Fixed author list in JOSS paper
