### 2019-06-21 0a6f8a7
[Bug #5] Fixed author list in JOSS paper
