### 2019-06-21 fee8183
[Enh #8] Added TEE-related specs to repository, and updated references in JOSS article.
