+++ docs/joss/paper.bib
@@ -1,9 +1,9 @@
 @Misc{TEESystemArch,
   author = {GP TEE community},
-  title =  {TEE System Architecture v1.2},
+  title =  {TEE System Architecture v1.0},
   year =   {2018},
   note =   {[Online; accessed 2019-05-06]},
-  url =    {https://globalplatform.org/specs-library/tee-system-architecture-v1-2/}
+  url =    {https://github.com/samsung/mtower/blob/master/docs/specs/gp/GPD_TEE_SystemArch_v1.0.pdf}
 }
 
 @Misc{ClientAPISpec,
@@ -11,15 +11,23 @@
   title =  {TEE Client API Specification v1.0},
   year =   {2010},
   note =   {[Online; accessed 2019-05-06]},
-  url =    {https://globalplatform.org/specs-library/tee-client-api-specification/}
+  url =    {https://github.com/samsung/mtower/blob/master/docs/specs/gp/TEE_Client_API_Specification-V1.0.pdf}
 }
 
 @Misc{CoreAPISpec,
   author = {GP TEE community},
-  title =  {TEE Internal Core API Specification v1.2},
+  title =  {TEE Internal Core API Specification v1.0},
   year =   {2018},
   note =   {[Online; accessed 2019-05-06]},
-  url =    {https://globalplatform.org/specs-library/tee-internal-core-api-specification-v1-2/}
+  url =    {https://github.com/samsung/mtower/blob/master/docs/specs/gp/GPD_TEE_Internal_API_Specification_v1.0.pdf}
+}
+
+@Misc{GPSpecLibrary,
+  author = {GP TEE community},
+  title =  {GP Specification Library},
+  year =   {2018},
+  note =   {[Online; accessed 2019-21-06]},
+  url =    {https://globalplatform.org/specs-library/}
 }
 
 
