### 2019-06-21 3ea96f5
[Enh #8] Added TEE-related specs to repository, and updated references in JOSS article.
