+++ docs/joss/paper.md
@@ -24,12 +24,14 @@ bibliography: paper.bib
 Embedded computing systems have already become omnipresent in day-to-day life, and in many cases these systems deal with confidential data and control access to valuable resources. Therefore, an ever-increasing effort is put into protecting these systems against malicious access. However, in many cases this effort lacks solid ground - the software and data being processed are being protected by outdated mechanisms, contain easy to exploit vulnerabilities and do not pose any real obstacles for determined and technically savvy adversary. Yet developing safe and efficient data protection mechanism from scratch is time-consuming and costly endeavor, and without independent assessment there's always risk of some undiscovered vulnerability that will render the mechanism useless. However, current generations of ARM-based microcontrollers offer a solid hardware foundation for building security mechanisms without having to design and implement basic security primitives - ARM TrustZone technology. Initially developed for ARM family of CPUs, TrustZone technology was later adopted into microcontroller (MCU) implementations of ARM architecture. Software libraries that implement security-related operations based on ARM TrustZone are readily available for Linux-based OSes and widely used, e.g., in smartphones running Android based on Linux kernel. However, these libraries are too cumbersome and resource-consuming to be directly adopted to embedded systems and MCU-based devices where CPU speeds are order of magnitude slower and RAM available for use is severely limited. mTower is an experimental standard-compliant implementation of GlobalPlatform Trusted Execution Environment (GP TEE) APIs based on ARM TrustZone technology. From the outset, it is designed to have minimal RAM footprint and avoid use of time-consuming operations that are undesirable in embedded systems. So far, the results seem promising, even though as of May 2019 only a partial implementation of GP TEE APIs is available in mTower.
 
 # Implementation overview
-In general, secure applications using TrustZone hardware as hardware security basis consist of two interacting components: Normal World (NW) and Secure World (SW). Normal World part is, in general, a standard Linux application that uses TEE library containing API functions to interact with Secure World part. Corresponding Secure World part of application is essentially a set of event handlers that are executed in hardware-protected area of RAM under control of specially designed operating system. Secure World part processes calls received from Normal World and deals with sensitive data that must be protected from unauthorized access (cryptographic keys, passwords, user identification data, etc.). Common operations performed by Secure World part of application include data encryption/decryption, user authentication, key generation, digital signing and signature verification. Overall architecture that applications should adhere to is described in TEE System Architecture v1.2 document [@TEESystemArch].
+In general, secure applications using TrustZone hardware as hardware security basis consist of two interacting components: Normal World (NW) and Secure World (SW). Normal World part is, in general, a standard Linux application that uses TEE library containing API functions to interact with Secure World part. Corresponding Secure World part of application is essentially a set of event handlers that are executed in hardware-protected area of RAM under control of specially designed operating system. Secure World part processes calls received from Normal World and deals with sensitive data that must be protected from unauthorized access (cryptographic keys, passwords, user identification data, etc.). Common operations performed by Secure World part of application include data encryption/decryption, user authentication, key generation, digital signing and signature verification. Overall architecture that applications should adhere to is described in TEE System Architecture v1.0 document [@TEESystemArch].
 
 mTower implementation follows this layout: it provides implementation of APIs described in GP TEE specification documents. Two most relevant TEE specifications are
 
 * TEE Client API Specification v1.0[@ClientAPISpec] describing the communication between NW applications and trusted applications residing in the SW.
-* TEE Internal Core API Specification v1.2[@CoreAPISpec] describing internal operations of SW parts of applications.
+* TEE Internal Core API Specification v1.0[@CoreAPISpec] describing internal operations of SW parts of applications.
+
[…]
