### 2019-06-21 39adf07
Removed author points from community-authored documents in paper.bib of JOSS article
