+++ docs/joss/paper.bib
@@ -1,5 +1,4 @@
 @Misc{TEESystemArch,
-  author = {GP TEE community},
   title =  {TEE System Architecture v1.0},
   year =   {2018},
   note =   {[Online; accessed 2019-05-06]},
@@ -7,7 +6,6 @@
 }
 
 @Misc{ClientAPISpec,
-  author = {GP TEE community},
   title =  {TEE Client API Specification v1.0},
   year =   {2010},
   note =   {[Online; accessed 2019-05-06]},
@@ -15,7 +13,6 @@
 }
 
 @Misc{CoreAPISpec,
-  author = {GP TEE community},
   title =  {TEE Internal Core API Specification v1.0},
   year =   {2018},
   note =   {[Online; accessed 2019-05-06]},
@@ -23,7 +20,6 @@
 }
 
 @Misc{GPSpecLibrary,
-  author = {GP TEE community},
   title =  {GP Specification Library},
   year =   {2018},
   note =   {[Online; accessed 2019-21-06]},
