### 2019-07-03 4840609
[Enh #6] Fixed some formatting
