### 2019-07-03 fd3e5d2
[Enh #6] Fixed some formatting
