### 2019-07-04 13e6e7a
[Enh #10] Added mTower functionality description for Nonsecure and Secure parts
