### 2019-07-04 b84e0aa
[Enh #10] Added mTower functionality description for Nonsecure and Secure parts
