+++ docs/mtower_functionality_description.md
@@ -1,0 +1,187 @@
+# mTower functionality description
+
+This file provides an overview of methods comprising mTower API available for use by apps. It contains 3 primary parts: mTower H/W API description, GP TEE API description, and mTower configuration options description.
+
+## 1. mTower H/W API
+mTower consists of Secure World and Normal World parts, which interact in a client-server pattern, with Normal World acting as a client and Secure World as a server. Normal World is an application that runs under FreeRTOS operating system. Below is a link to description of FreeRTOS API:
+[The FreeRTOS Reference Manual](https://www.freertos.org/Documentation/FreeRTOS_Reference_Manual_V10.0.0.pdf)
+SecureWorld is basically just a set of event handlers, and its API is completely defined by developers themselves. Example of API functions is given in GP TEE Internal API table.
+
+## 2. mTower and GP TEE API descriptions
+Table below gives a listing of API functions for Normal World. For a detailed description of each function, please see [TEE Client API](specs/gp/TEE_Client_API_Specification-V1.0.pdf).
+
+| **GP TEE Client API**           |**Implemented**| **Comments**              |
+|---------------------------------|-----|-------------------------------------|
+| TEEC_InitializeContext          | Yes |                                     |
+| TEEC_FinalizeContext            | Yes |                                     |
+| TEEC_RegisterSharedMemory       | No  |                                     |
+| TEEC_AllocateSharedMemory       | Yes |                                     |
+| TEEC_ReleaseSharedMemory        | Yes |                                     |
+| TEEC_OpenSession                | Yes |                                     |
+| TEEC_CloseSession               | Yes |                                     |
+| TEEC_InvokeCommand              | Yes |                                     |
+| TEEC_RequestCancellation        | No  |                                     |
+| TEEC_PARAM_TYPES                | Yes |                                     |
+
+A list of Secure World API functions is given in a table below.For a detailed description of each function, please see [TEE Internal API](specs/gp/GPD_TEE_Internal_API_Specification_v1.0.pdf).
+
+| **GP TEE Internal API**         |**Implemented**| **Comments**              |
+|---------------------------------|-----|-------------------------------------|
+| **Trusted Core Framework**      |     |                                     |
+| **TA Interface**                |     |                                     |
+| TA_CreateEntryPoint             | Yes | Every TA includes this function replacing <TA> with <TA_name> prefix |
+| TA_DestroyEntryPoint            | Yes | Every TA includes this function replacing <TA> with <TA_name> prefix |
+| TA_OpenSessionEntryPoint        | Yes | Every TA includes this function replacing <TA> with <TA_name> prefix |
+| TA_CloseSessionEntryPoint       | Yes | Every TA includes this function replacing <TA> with <TA_name> prefix |
+| TA_InvokeCommandEntryPoint      | Yes | Every TA includes this function replacing <TA> with <TA_name> prefix |
+| **Property Access Functions**   |     |                                     |
+| TEE_GetPropertyAsString         | No  |                                     |
+| TEE_GetPropertyAsBool           | No  |                                     |
+| TEE_GetPropertyAsU32            | No  |                                     |
+| TEE_GetPropertyAsBinaryBlock    | No  |                                     |
+| TEE_GetPropertyAsUUID           | No  |                                     |
+| TEE_GetPropertyAsIdentity       | No  |                                     |
+| TEE_AllocatePropertyEnumerator  | No  |                                     |
+| TEE_FreePropertyEnumerator      | No  |                                     |
+| TEE_StartPropertyEnumerator     | No  |                                     |
+| TEE_ResetPropertyEnumerator     | No  |                                     |
+| TEE_GetPropertyName             | No  |                                     |
+| TEE_GetNextProperty             | No  |                                     |
[…]
