### 2019-07-04 383711b
[Enh #7] Added guide on creating mTower-based app
