### 2019-07-04 46a2535
[Enh #7] Added guide on creating mTower-based app
