### 2019-07-05 81f72f5
[Enh #9] Added more detailed explanation of "how to add a new platform"
