+++ docs/port-new-platform.md
@@ -3,8 +3,8 @@
 # Contents
 1. [Introduction](#1-introduction)
 2. [Add a new platform](#2-add-a-new-platform)
-    1. [arch/<newdev>](#2-1-arch-newdev)
-    2. [config/<newdev>](#2-2-config-newdev)
+    2.1 [arch/<newdev>](#2-1-arch-newdev)
+    2.2 [config/<newdev>](#2-2-config-newdev)
 3. [Code inclusion to mTower upstream](#3-code-inclusion-to-mtower-upstream)
 
 ## 1. Introduction
@@ -19,18 +19,111 @@ configuring UART, memory allocator etc. For simplicity let us call our new
 platform for "newdev" just so we have something to refer to when writing
 examples further down.
 
-### 2.1 arch/<newdev>
+### 2.1 arch/newdev
 
-In `arch/` you will find all the currently supported architecture. That is where you
+In `arch/` you will find all the currently supported architectures. That is where you
 are supposed to add a new platform or modify an existing one. Typically you will
 find this set of files in a specific platform folder. Clone from an existing
 platform.
 
-> TBD
+#### 2.1.1 Adding new architecture
+If a new platform uses an architecture not included into build before (as of now, only cortex-m23 architecture is included), you need to add a new block into `arch/Kconfig` file that specifies the name of new architecture. For example:
 
-### 2.2 config/<newdev>
+```
+
+config ARCH_CORTEX_M33
+	bool "cortex-m33"
+	---help---
+		The ARM architecture
+
+```
+
+Also, under line
+
+```
+default "cortex-m23"    if ARCH_CORTEX_M23`
+```
+
+ you need to add a new line for architecture that is added, e. g.:
+ 
+```
+default "cortex-m33"    if ARCH_CORTEX_M33`
+```
+
+Under
+
+```
+if ARCH_CORTEX_M23
+source arch/cortex-m23/Kconfig
+endif
+```
+block you need to add a similar block for new architecture, e.g.,
+
+```
+if ARCH_CORTEX_M33
+source arch/cortex-m33/Kconfig
+endif
+```
+
+
+#### 2.1.2 Adding chip family
+If a new platform uses chip family not included into build before, add a new folder into architecture folder, e.g.,
+`arch/cortex-m23/new_chip_family`.
+After that, add a new chip family into `Kconfig` file, e.g., `arch/cortex-m23/Kconfig`:
+
+```
+config ARCH_FAMILY_NEW_CHIP_FAMILY
+	bool "NEW_CHIP_FAMILY"
+	---help---
+		The new chip family
+
+endchoice
+
+config ARCH_FAMILY
+	string
+	default "m2351"            if ARCH_FAMILY_M2351
+	default "NEW_CHIP_FAMILY" if ARCH_FAMILY_NEW_CHIP_FAMILY
+
+if ARCH_FAMILY_M2351
+comment "M2351 Configuration Options"
+source arch/cortex-m23/m2351/Kconfig
+endif
+
+if ARCH_FAMILY_NEW_CHIP_FAMILY
+comment "NEW_CHIP_FAMILY Configuration Options"
+source arch/cortex-m23/new_chip_family/Kconfig
+endif
+
+```
+
+#### 2.1.3 Adding platform
+Last step (and the only one that is always needed) is adding new platform folder and updating corresponding `KConfig` file. Add a folder named after new platform into chip family folder, e.g., `arch/cortex-m23/new_chip_family/new_platform`. Then open `Kconfig` file in this new folder, and add the following lines into it:
+
+```
+config NEW_PLATFORM
+	bool "new_platform board"
+	---help---
+		new platform board.
+
+endchoice
[…]
