### 2019-07-05 8089bd0
[Enh #9] Added more detailed explanation of "how to add a new platform"
