+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/nonsecure/warn_err.c
@@ -69,6 +69,6 @@ void errx(int eval, const char *fmt, ...)
   fflush(stdout);
   va_end(va);
 
-  vTaskDelete(NULL);
+//  vTaskDelete(NULL);
 }
 
