### 2019-07-18 9a89f31
Updated article for JOSS
