+++ docs/joss/paper.bib
@@ -1,22 +1,64 @@
+@Misc{ZWave,
+  title =  {ZWave lock vulnerability},
+  year =   {2014},
+  note =   {[Online; accessed 2019-07-02]},
+  url =    {https://suretycam.com/can-hackers-unlock-my-z-wave-door-lock}
+}
+
+@Misc{Forbes,
+  title =  {Insulin pump vulnerability},
+  year =   {2014},
+  note =   {[Online; accessed 2019-07-02]},
+  url =    {https://www.forbes.com/sites/ericbasu/2013/08/03/hacking-insulin-pumps-and-other-medical-devices-reality-not-fiction}
+}
+
+@Misc{IOactive,
+  title =  {Automotive devices vulnerabilities},
+  year =   {2014},
+  note =   {[Online; accessed 2019-07-02]},
+  url =    {http://www.ioactive.com/pdfs/IOActive_Adventures_in_Automotive_Networks_and_Control_Units.pdf}
+}
+
+@Misc{Kinibi-M,
+  title =  {Trustonic Kinibi-M},
+  year =   {2017},
+  note =   {[Online; accessed 2019-07-02]},
+  url =    {https://www.trustonic.com/news/blog/not-just-droning-rise-kinibi-m/}
+}
+
+@Misc{ProvenCore-M,
+  title =  {ProvenCore-M},
+  year =   {2018},
+  note =   {[Online; accessed 2019-07-02]},
+  url =    {http://www.provenrun.com/products/provencore-m/}
+}
+
+@Misc{CoreLockr-TZ,
+  title =  {CoreLockr-TZ},
+  year =   {2018},
+  note =   {[Online; accessed 2019-07-02]},
+  url =    {https://www.sequiturlabs.com/corelockrtz/}
+}
+
 @Misc{TEESystemArch,
   title =  {TEE System Architecture v1.0},
   year =   {2018},
   note =   {[Online; accessed 2019-05-06]},
-  url =    {https://github.com/samsung/mtower/blob/master/docs/specs/gp/GPD_TEE_SystemArch_v1.0.pdf}
+  url =    {https://globalplatform.org/specs-library/}
 }
 
 @Misc{ClientAPISpec,
   title =  {TEE Client API Specification v1.0},
   year =   {2010},
   note =   {[Online; accessed 2019-05-06]},
-  url =    {https://github.com/samsung/mtower/blob/master/docs/specs/gp/TEE_Client_API_Specification-V1.0.pdf}
+  url =    {https://globalplatform.org/specs-library/}
 }
 
 @Misc{CoreAPISpec,
   title =  {TEE Internal Core API Specification v1.0},
   year =   {2018},
   note =   {[Online; accessed 2019-05-06]},
-  url =    {https://github.com/samsung/mtower/blob/master/docs/specs/gp/GPD_TEE_Internal_API_Specification_v1.0.pdf}
+  url =    {https://globalplatform.org/specs-library}
 }
 
 @Misc{GPSpecLibrary,
