+++ docs/joss/paper.md
@@ -21,36 +21,49 @@ bibliography: paper.bib
 ---
 
 # Background
-Embedded computing systems have already become omnipresent in day-to-day life, and in many cases these systems deal with confidential data and control access to valuable resources. Therefore, an ever-increasing effort is put into protecting these systems against malicious access. However, in many cases this effort lacks solid ground - the software and data being processed are being protected by outdated mechanisms, contain easy to exploit vulnerabilities and do not pose any real obstacles for determined and technically savvy adversary. Yet developing safe and efficient data protection mechanism from scratch is time-consuming and costly endeavor, and without independent assessment there's always risk of some undiscovered vulnerability that will render the mechanism useless. However, current generations of ARM-based microcontrollers offer a solid hardware foundation for building security mechanisms without having to design and implement basic security primitives - ARM TrustZone technology. Initially developed for ARM family of CPUs, TrustZone technology was later adopted into microcontroller (MCU) implementations of ARM architecture. Software libraries that implement security-related operations based on ARM TrustZone are readily available for Linux-based OSes and widely used, e.g., in smartphones running Android based on Linux kernel. However, these libraries are too cumbersome and resource-consuming to be directly adopted to embedded systems and MCU-based devices where CPU speeds are order of magnitude slower and RAM available for use is severely limited. mTower is an experimental standard-compliant implementation of GlobalPlatform Trusted Execution Environment (GP TEE) APIs based on ARM TrustZone technology. From the outset, it is designed to have minimal RAM footprint and avoid use of time-consuming operations that are undesirable in embedded systems. So far, the results seem promising, even though as of May 2019 only a partial implementation of GP TEE APIs is available in mTower.
+Embedded computing systems have already become ubiquitous in daily life. In many cases these systems deal with confidential data and control access to valuable resources. Significant efforts are put into securing these systems against malicious access. However the software and data are often protected by outdated mechanisms that contain well-known vulnerabilities, thus do not pose any real obstacles for technically savvy adversary. Consequences of deliberate attack against embedded systems can be severe, e.g.:
+
+* poorly implemented authorization/authentication mechanism allows hackers to unlock a smart door lock [@ZWave]
+* poor data protection in network-connected medical device can pose life risk [@Forbes]
+* hardcoded cryptographic keys and application vulnerable to reverse engineering in automotive platform[@IOactive]
+
[…]
