### 2019-07-18 #1494 @gkthiruvathukal
@omoliavko I think that is all we need to move toward acceptance. Please let all of us (me and the reviewers) know when the test sequence/scenarios and documentation are ready.
