### 2019-07-23 #1494 @gkthiruvathukal
@federeghe @gradvohl Can you let me know if you have any more issues? I'm going to assume you are ok, given that @omoliavko has addressed our last point of feedback.

So here is the final checklist for @omoliavko to fill in. Can you please do the following? (This is something I am doing for all submissions now, so you can just check the ones you already did above.)

Once these are done, I will do a final proofread and signal the JOSS EICs to process acceptance.
