### 2019-07-23 #1494 @gradvohl
Good Morning,

The authors addressed most of the issues I pointed. However, I still find some copyrighted code, such as ``tee_client/public/teec_trace.h`` and ``mTower/tee_client/libteec/src/tee_client_api.c``. Is it a problem for JOSS?
Regarding the other issues, I consider that the authors addressed them.
Best regards.
