### 2019-07-23 edbb7e1
Fixed name function inacuracy in hello_world_ta.c
