### 2019-07-23 4525342
Refactored some source code; clarified licensing terms in COPYING file
