### 2019-07-23 #1494 @gkthiruvathukal
@gradvohl This would be a problem for JOSS, especially if the code does not belong to the authors. Even if it does, the licensing needs to be clear.

@omoliavko Can you please clarify the status of the tee_client code?
