### 2019-07-24 #1494 @gkthiruvathukal
@omoliavko Thank you for the update. My last question is whether this is clearly documented somewhere in your code. Are the licenses for these codes included in the folder containing them?
