### 2019-07-24 #1494 @gkthiruvathukal
@omoliavko So can you go to my earlier checklist and check off all of the items? I believe we are at a point where you can prepare the Zenodo archive and tag the current release.
