### 2019-07-25 #1494 @gkthiruvathukal
@omoliavko I believe all issues have been addressed thus far. Let's do this. If  @gradvohl and @federeghe have any comments, I'd like to receive these by end of week. If there are no objections by the weekend (any time on earth) I will ask you to tag the release and create the Zenodo archive. We should be able to wrap up (and accept) this submission soon. Thanks for your patience and diligence to work on all of the issues raised during review.
