### 2019-08-12 712a76e
Added H/W security exception example
