+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/nonsecure/main_ns.c
@@ -498,6 +498,98 @@ void vApplicationTickHook( void )
   ;
 }
 
+#ifdef CONFIG_APPS_HW_SECURITY_EXCEPTION_EXAMPLE
+void menu_security_exception_example_ns(void)
+{
+  char ch;
+  int temp = 1;
+
+  printf("+---------------------------------------------------------------------+\n");
+  printf("| Type  | Range addresses         | Size      | Remarks               |\n");
+  printf("+---------------------------------------------------------------------+\n");
+  printf("| FLASH | 0x10040000 - 0x0007FFFF | 0x40000   | 256k                  |\n");
+  printf("| RAM   | 0x30008000 - 0x30017FFF | 0x10000   | 64k                   |\n");
+  printf("| GPIO  | PC                      |           | Port C is non-secure  |\n");
+  printf("| GPIO  | PA,PB                   |           | Port A,B are secure   |\n");
+  printf("+---------------------------------------------------------------------+\n\n");
+
+  printf("+---------------------------------------------------------------------+\n");
+  printf("| Key | Action                                   | Expected response  |\n");
+  printf("+---------------------------------------------------------------------+\n");
+  printf("| [1] | Read SRAM secure address 0x30000000      | Hard fault         |\n");
+  printf("| [2] | Read SRAM secure address 0x20000000      | Hard fault         |\n");
+  printf("| [3] | Read SRAM non-secure address 0x%08X  | Access successful  |\n",&temp);
+  printf("|     | Read SRAM non-secure address 0x%08X  | Hard fault         |\n",(~0x10000000 & (unsigned int)&temp));
+  printf("| [4] | Read FLASH non-secure address 0x00000000 | Hard fault         |\n");
+  printf("| [5] | Read FLASH non-secure address 0x10000000 | Hard fault         |\n");
+  printf("| [6] | Read FLASH secure address 0x10040000     | Access successful  |\n");
+  printf("|     | Read FLASH secure address 0x00040000     | Hard fault         |\n");
+  printf("| [7] | Read GPIO non-secure port PC1_NS         | Access successful  |\n");
+  printf("|     | Write 0 GPIO non-secure port by PC1_NS   | Access successful  |\n");
+  printf("|     | Write 1 GPIO non-secure port by PC1      | Hard fault         |\n");
+  printf("| [8] | Read GPIO secure port PA10_NS            | RAZWI, sec. violat.|\n");
+  printf("|     | Write 0 GPIO secure port by PA10_NS      | RAZWI, sec. violat.|\n");
+  printf("|     | Write 1 GPIO secure port by PA10         | Hard fault         |\n");
+  printf("| [9] | Write 0 to address 0x0 (directly)        | Hard fault         |\n");
+  printf("| [a] | Read 0x30018000 address (nonexistent)    | Hard fault         |\n");
+  printf("+---------------------------------------------------------------------+\n");
+
+  printf("\n[%c] ", ch = getchar());
+
+  switch (ch) {
+    case '1':
+      printf("Read SRAM secure address 0x30000000: Hard fault occurs in secure\n");
+      M32(0x30000000);
+      break;
+    case '2':
+      printf("Read SRAM secure address 0x20000000: Hard fault occurs in secure\n");
+      M32(0x20000000);
+      break;
+    case '3':
+      printf("Read SRAM non-secure address 0x%08X = %08x\n", &temp, M32(&temp));
[…]
