+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/secure/main.c
@@ -24,9 +24,20 @@
 #include <stdio.h>
 #include "M2351.h"
 #include "partition_M2351.h"
+#include "config.h"
 #include "version.h"
 
 /* Pre-processor Definitions. */
+#define NORMAL  "\033[0m"
+#define BLACK   "\033[0;30m1"
+#define RED     "\033[0;31m"
+#define GREEN   "\033[0;32m"
+#define YELLOW  "\033[0;33m"
+#define BLUE    "\033[0;34m"
+#define MAGENTA "\033[0;35m"
+#define CYAN    "\033[0;36m"
+#define GRAY    "\033[0;37m"
+
 /** Start address for non-secure boot image */
 #define NEXT_BOOT_BASE  0x10040000
 /** Instruction Code of "B ."  */
@@ -287,6 +298,138 @@ int32_t LED_Off(void)
   return 1;
 }
 
+#ifdef CONFIG_APPS_HW_SECURITY_EXCEPTION_EXAMPLE
+void hw_security_exception_init(void)
+{
+  /* Init PC for Nonsecure LED control */
+  GPIO_SetMode(PC_NS, BIT1, GPIO_MODE_OUTPUT);
+
+  /* Set Hard Fault to secure */
+//  SCB->AIRCR = (0x05FA << 16) | ((SCB->AIRCR&0xfffful) & (~(1 << SCB_AIRCR_BFHFNMINS_Pos)));
+//  SCB->AIRCR = (0x05FA << 16) | ((SCB->AIRCR & 0xfffful) | (1 << SCB_AIRCR_BFHFNMINS_Pos));
+
+  if (SCB->AIRCR & SCB_AIRCR_BFHFNMINS_Msk) {
+    printf("Non-secure Hard Fault handled by Non-secure code.\n");
+  } else {
+    printf("Non-secure Hard Fault handled by secure code.\n");
+  }
+//  printf("SCB->AIRCR = 0x%08x\n", SCB->AIRCR);
+
+  NVIC_EnableIRQ(SCU_IRQn);
+}
+void menu_security_exception_example(void)
+{
+  char ch;
+  int temp = 1;
+
+//  printf("SCU->SVIOIEN = 0x%08x\n",SCU->SVIOIEN);
+//  printf("SCU->SRAMNSSET = 0x%08x\n",SCU->SRAMNSSET);
+
+  printf("+---------------------------------------------------------------------+\n");
+  printf("| Type  | Range addresses         | Size      | Remarks               |\n");
+  printf("+---------------------------------------------------------------------+\n");
+  printf("| FLASH | 0x00000000 - 0x0003F000 | 0x3F000   | 252k                  |\n");
+  printf("| RAM   | 0x20000000 - 0x20007FFF | 0x8000    | 32k                   |\n");
+  printf("| NCS   | 0x0003F000 - 0x0003FFFF | 0x1000    | 4k                    |\n");
+  printf("| GPIO  | PC                      |           | Port C is non-secure  |\n");
+  printf("| GPIO  | PA,PB                   |           | Port A,B are secure   |\n");
+  printf("+---------------------------------------------------------------------+\n\n");
+
+  printf("+---------------------------------------------------------------------+\n");
+  printf("| Key | Action                                   | Expected response  |\n");
+  printf("+---------------------------------------------------------------------+\n");
+  printf("| [1] | Read SRAM non-secure address 0x30017000  | Access successful  |\n");
+  printf("|     | Read SRAM non-secure address 0x20017000  | RAZWI              |\n");
+  printf("| [2] | Read SRAM secure address 0x%08X      | Access successful  |\n",&temp);
[…]

[…]
