### 2019-08-12 1d84a3b
Added info required for progressing from v0.1 to v0.2
