+++ README.md
@@ -37,7 +37,7 @@ between platforms, a `PLATFORM` flag has been introduced.
 
 | Platform              | Composite PLATFORM flag            | Maintained |
 |-----------------------|------------------------------------|------------|
-| [NuMaker-PFM-M2351]   |`PLATFORM=numaker_pfm_m2351`    | v0.1       |
+| [NuMaker-PFM-M2351]   |`PLATFORM=numaker_pfm_m2351`    | v0.2       |
 
 Fo information on adding a new platform see the [how to add a platform].
 
