### 2019-08-15 befa9d5
Added GPIO Secure Attribution Configuration
