### 2019-08-15 8d0d4c1
Added Peripheral Secure Attribution Configuration
