+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/partition_M2351.h
@@ -87,118 +87,6 @@ __STATIC_INLINE void FMC_NSBA_Setup(void)
 // <h> Peripheral Secure Attribution Configuration
 */
 
-/*
-    PNSSET0
-*/
-/*
-// Module 0..31
-//   <o.9>  USBH       <0=> Secure <1=> Non-Secure
-//   <o.13>  SD0   <0=> Secure <1=> Non-Secure
-//   <o.16>  EBI    <0=> Secure <1=> Non-Secure
-//   <o.24>  PDMA1      <0=> Secure <1=> Non-Secure
-*/
-#define SCU_INIT_PNSSET0_VAL      0x0
-/*
-    PNSSET1
-*/
-/*
-// Module 0..31
-//   <o.17>  CRC       <0=> Secure <1=> Non-Secure
-//   <o.18>  CRPT   <0=> Secure <1=> Non-Secure
-*/
-#define SCU_INIT_PNSSET1_VAL      0x0
-/*
-    PNSSET2
-*/
-/*
-// Module 0..31
-//   <o.1>  RTC       <0=> Secure <1=> Non-Secure
-//   <o.3>  EADC   <0=> Secure <1=> Non-Secure
-//   <o.5>  ACMP01     <0=> Secure <1=> Non-Secure
-//
-//   <o.7>  DAC      <0=> Secure <1=> Non-Secure
-//   <o.8>  I2S0      <0=> Secure <1=> Non-Secure
-//   <o.13>  OTG      <0=> Secure <1=> Non-Secure
-//   <o.17>  TMR23      <0=> Secure <1=> Non-Secure
-//   <h> EPWM
-//   <o.24>  EPWM0      <0=> Secure <1=> Non-Secure
-//   <o.25>  EPWM1      <0=> Secure <1=> Non-Secure
-//   <o.26>  BPWM0      <0=> Secure <1=> Non-Secure
-//   <o.27>  BPWM1      <0=> Secure <1=> Non-Secure
-//   </h>
-*/
-#define SCU_INIT_PNSSET2_VAL      0x0
-/*
-    PNSSET3
-*/
-/*
-// Module 0..31
-//   <h>  SPI
-//   <o.0>  QSPI0       <0=> Secure <1=> Non-Secure
-//   <o.1>  SPI0   <0=> Secure <1=> Non-Secure
-//   <o.2>  SPI1      <0=> Secure <1=> Non-Secure
-//   <o.3>  SPI2    <0=> Secure <1=> Non-Secure
-//   <o.4>  SPI3      <0=> Secure <1=> Non-Secure
-//   </h>
-//   <h> UART
-//   <o.16>  UART0      <0=> Secure <1=> Non-Secure
-//   <o.17>  UART1      <0=> Secure <1=> Non-Secure
-//   <o.18>  UART2      <0=> Secure <1=> Non-Secure
-//   <o.19>  UART3      <0=> Secure <1=> Non-Secure
-//   <o.20>  UART4      <0=> Secure <1=> Non-Secure
-//   <o.21>  UART5      <0=> Secure <1=> Non-Secure
-//   </h>
-*/
-#define SCU_INIT_PNSSET3_VAL      0x20000
-/*
-    PNSSET4
-*/
-/*
-// Module 0..31
-//   <h> I2C
[…]
