### 2019-08-16 2abab5e
Added setting up configuration of Secure/Non-Secure region
