+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/partition_M2351.h
@@ -350,211 +350,6 @@ __STATIC_INLINE void SCU_Setup(void)
 // </e>
 */
 
-
-/*
-// <h>Enable and Set Secure/Non-Secure region
-*/
-#define SAU_REGIONS_MAX   8                 /* Max. number of SAU regions */
-
-/*
-//   <e>SAU Region 0
-//   <i> Setup SAU Region 0
-*/
-#define SAU_INIT_REGION0    1 //0
-/*
-//     <o>Start Address <0-0xFFFFFFE0>
-*/
-#define SAU_INIT_START0     0x20000000 //0x0003F000      /* start address of SAU region 0 */
-/*
-//     <o>End Address <0x1F-0xFFFFFFFF>
-*/
-#define SAU_INIT_END0       0x20008000 //0x0003FFFF      /* end address of SAU region 0 */
-/*
-//     <o>Region is
-//         <0=>Non-Secure
-//         <1=>Secure, Non-Secure Callable
-*/
-#define SAU_INIT_NSC0       1
-/*
-//   </e>
-*/
-
-/*
-//   <e>SAU Region 1
-//   <i> Setup SAU Region 1
-*/
-#define SAU_INIT_REGION1    0
-/*
-//     <o>Start Address <0-0xFFFFFFE0>
-*/
-#define SAU_INIT_START1     0x10040000
-/*
-//     <o>End Address <0x1F-0xFFFFFFFF>
-*/
-#define SAU_INIT_END1       0x1007FFFF
-/*
-//     <o>Region is
-//         <0=>Non-Secure
-//         <1=>Secure, Non-Secure Callable
-*/
-#define SAU_INIT_NSC1       0
-/*
-//   </e>
-*/
-
-/*
-//   <e>SAU Region 2
-//   <i> Setup SAU Region 2
-*/
-#define SAU_INIT_REGION2    0
-/*
-//     <o>Start Address <0-0xFFFFFFE0>
-*/
-#define SAU_INIT_START2     0x2000F000
-/*
-//     <o>End Address <0x1F-0xFFFFFFFF>
-*/
-#define SAU_INIT_END2       0x2000FFFF
-/*
-//     <o>Region is
-//         <0=>Non-Secure
-//         <1=>Secure, Non-Secure Callable
-*/
-#define SAU_INIT_NSC2       1
-/*
-//   </e>
-*/
-
-/*
-//   <e>SAU Region 3
-//   <i> Setup SAU Region 3
-*/
-#define SAU_INIT_REGION3    1
-/*
-//     <o>Start Address <0-0xFFFFFFE0>
-*/
-#define SAU_INIT_START3     0x3f000
-/*
-//     <o>End Address <0x1F-0xFFFFFFFF>
-*/
-#define SAU_INIT_END3       0x3f7ff
-/*
-//     <o>Region is
-//         <0=>Non-Secure
-//         <1=>Secure, Non-Secure Callable
-*/
-#define SAU_INIT_NSC3       1
-/*
-//   </e>
-*/
-
-/*
-   <e>SAU Region 4
-   <i> Setup SAU Region 4
-*/
-#define SAU_INIT_REGION4    1
-/*
-     <o>Start Address <0-0xFFFFFFE0>
-*/
-#define SAU_INIT_START4     FMC_NON_SECURE_BASE      /* start address of SAU region 4 */
-
-/*
-     <o>End Address <0x1F-0xFFFFFFFF>
-*/
[…]
