+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/secure/main.c
@@ -416,20 +416,102 @@ void HardFault_Handler(void)
 void SCU_IRQHandler(void)
 {
   printf(RED "SCU_IRQHandler detected: ");
+#ifdef CONFIG_APB0IEN_SECURE_VIOLATION_INTERRUPT
+  if((SCU->SVINTSTS & SCU_SVIOIEN_APB0IEN_Msk) == SCU_SVIOIEN_APB0IEN_Msk) {
+    SCU->SVINTSTS |= SCU_SVIOIEN_APB0IEN_Msk;
+    printf("APB0 violation interrupt event\n");
+}
+#endif
+#ifdef CONFIG_APB1IEN_SECURE_VIOLATION_INTERRUPT
+  if((SCU->SVINTSTS & SCU_SVIOIEN_APB1IEN_Msk) == SCU_SVIOIEN_APB1IEN_Msk) {
+    SCU->SVINTSTS |= SCU_SVIOIEN_APB1IEN_Msk;
+    printf("APB1 violation interrupt event\n");
+}
+#endif
+#ifdef CONFIG_GPIOIEN_SECURE_VIOLATION_INTERRUPT
   if((SCU->SVINTSTS & SCU_SVIOIEN_GPIOIEN_Msk) == SCU_SVIOIEN_GPIOIEN_Msk) {
     SCU->SVINTSTS |= SCU_SVIOIEN_GPIOIEN_Msk;
     printf("GPIO violation interrupt event\n");
-  }
-//  TBD
-//  if((SCU->SVINTSTS & SCU_SVIOIEN_SRAM0IEN_Msk) == SCU_SVIOIEN_SRAM0IEN_Msk) {
-//    SCU->SVINTSTS |= SCU_SVIOIEN_SRAM0IEN_Msk;
-//    printf("SRAM0 violation interrupt event\n");
-//  }
-//  if((SCU->SVINTSTS & SCU_SVIOIEN_SRAM1IEN_Msk) == SCU_SVIOIEN_SRAM1IEN_Msk) {
-//    SCU->SVINTSTS |= SCU_SVIOIEN_SRAM1IEN_Msk;
-//    printf("SRAM1 violation interrupt event\n");
-//  }
-  printf(NORMAL);
+}
+#endif
+#ifdef CONFIG_EBIIEN_SECURE_VIOLATION_INTERRUPT
+  if((SCU->SVINTSTS & SCU_SVIOIEN_EBIIEN_Msk) == SCU_SVIOIEN_EBIIEN_Msk) {
+    SCU->SVINTSTS |= SCU_SVIOIEN_EBIIEN_Msk;
+    printf("EBI violation interrupt event\n");
+}
+#endif
+#ifdef CONFIG_USBHIEN_SECURE_VIOLATION_INTERRUPT
+  if((SCU->SVINTSTS & SCU_SVIOIEN_USBHIEN_Msk) == SCU_SVIOIEN_USBHIEN_Msk) {
+    SCU->SVINTSTS |= SCU_SVIOIEN_USBHIEN_Msk;
+    printf("USBH violation interrupt event\n");
+}
+#endif
+#ifdef CONFIG_CRCIEN_SECURE_VIOLATION_INTERRUPT
+  if((SCU->SVINTSTS & SCU_SVIOIEN_CRCIEN_Msk) == SCU_SVIOIEN_CRCIEN_Msk) {
+    SCU->SVINTSTS |= SCU_SVIOIEN_CRCIEN_Msk;
+    printf("CRC violation interrupt event\n");
+}
+#endif
[…]
