+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/partition_M2351.h
@@ -318,37 +318,71 @@ __STATIC_INLINE void SCU_Setup(void)
     SCU->SRAMNSSET |= (1U << i);
   }
 
-  SCU->SVIOIEN = SCU_SVIOIEN_GPIOIEN_Msk;
-//    | SCU_SVIOIEN_SRAM0IEN_Msk | SCU_SVIOIEN_SRAM1IEN_Msk;
+#ifdef CONFIG_APB0IEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_APB0IEN_Msk);
+#endif
+
+#ifdef CONFIG_APB1IEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_APB1IEN_Msk);
+#endif
+
+#ifdef CONFIG_GPIOIEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_GPIOIEN_Msk);
+#endif
+
+#ifdef CONFIG_EBIIEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_EBIIEN_Msk);
+#endif
+
+#ifdef CONFIG_USBHIEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_USBHIEN_Msk);
+#endif
+
+#ifdef CONFIG_CRCIEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_CRCIEN_Msk);
+#endif
+
+#ifdef CONFIG_SDH0IEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_SDH0IEN_Msk);
+#endif
+
+#ifdef CONFIG_PDMA0IEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_PDMA0IEN_Msk);
+#endif
+
+#ifdef CONFIG_PDMA1IEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_PDMA1IEN_Msk);
+#endif
+
+#ifdef CONFIG_SRAM0IEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_SRAM0IEN_Msk);
+#endif
+
+#ifdef CONFIG_SRAM1IEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_SRAM1IEN_Msk);
+#endif
+
+#ifdef CONFIG_FMCIEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_FMCIEN_Msk);
+#endif
+
+#ifdef CONFIG_FLASHIEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_FLASHIEN_Msk);
+#endif
+
+#ifdef CONFIG_SCUIEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_SCUIEN_Msk);
+#endif
+
+#ifdef CONFIG_SYSIEN_SECURE_VIOLATION_INTERRUPT
+  SCU_ENABLE_INT(SCU_SVIOIEN_SYSIEN_Msk);
+#endif
+
+#ifdef CONFIG_CRPTIEN_SECURE_VIOLATION_INTERRUPT
[…]
