### 2019-08-20 d861801
Fixed some inaccuracies
