+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/secure/main.c
@@ -408,7 +408,7 @@ void menu_security_exception_example(void)
 
 void HardFault_Handler(void)
 {
-  printf(RED "Secure HardFault_Handler: invalid memory access or malware activity detected\n" NORMAL);
+  printf(RED "Secure Hard Fault Handler: invalid memory access or malware activity detected\n" NORMAL);
 
   while(1);
 }
@@ -419,99 +419,100 @@ void SCU_IRQHandler(void)
 #ifdef CONFIG_APB0IEN_SECURE_VIOLATION_INTERRUPT
   if((SCU->SVINTSTS & SCU_SVIOIEN_APB0IEN_Msk) == SCU_SVIOIEN_APB0IEN_Msk) {
     SCU->SVINTSTS |= SCU_SVIOIEN_APB0IEN_Msk;
-    printf("APB0 violation interrupt event\n");
+    printf("APB0 ");
 }
 #endif
 #ifdef CONFIG_APB1IEN_SECURE_VIOLATION_INTERRUPT
   if((SCU->SVINTSTS & SCU_SVIOIEN_APB1IEN_Msk) == SCU_SVIOIEN_APB1IEN_Msk) {
     SCU->SVINTSTS |= SCU_SVIOIEN_APB1IEN_Msk;
-    printf("APB1 violation interrupt event\n");
+    printf("APB1 ");
 }
 #endif
 #ifdef CONFIG_GPIOIEN_SECURE_VIOLATION_INTERRUPT
   if((SCU->SVINTSTS & SCU_SVIOIEN_GPIOIEN_Msk) == SCU_SVIOIEN_GPIOIEN_Msk) {
     SCU->SVINTSTS |= SCU_SVIOIEN_GPIOIEN_Msk;
-    printf("GPIO violation interrupt event\n");
+    printf("GPIO ");
 }
 #endif
 #ifdef CONFIG_EBIIEN_SECURE_VIOLATION_INTERRUPT
   if((SCU->SVINTSTS & SCU_SVIOIEN_EBIIEN_Msk) == SCU_SVIOIEN_EBIIEN_Msk) {
     SCU->SVINTSTS |= SCU_SVIOIEN_EBIIEN_Msk;
-    printf("EBI violation interrupt event\n");
+    printf("EBI ");
 }
 #endif
 #ifdef CONFIG_USBHIEN_SECURE_VIOLATION_INTERRUPT
   if((SCU->SVINTSTS & SCU_SVIOIEN_USBHIEN_Msk) == SCU_SVIOIEN_USBHIEN_Msk) {
     SCU->SVINTSTS |= SCU_SVIOIEN_USBHIEN_Msk;
-    printf("USBH violation interrupt event\n");
+    printf("USBH ");
 }
 #endif
 #ifdef CONFIG_CRCIEN_SECURE_VIOLATION_INTERRUPT
   if((SCU->SVINTSTS & SCU_SVIOIEN_CRCIEN_Msk) == SCU_SVIOIEN_CRCIEN_Msk) {
     SCU->SVINTSTS |= SCU_SVIOIEN_CRCIEN_Msk;
-    printf("CRC violation interrupt event\n");
+    printf("CRC ");
 }
 #endif
 #ifdef CONFIG_SDH0IEN_SECURE_VIOLATION_INTERRUPT
   if((SCU->SVINTSTS & SCU_SVIOIEN_SDH0IEN_Msk) == SCU_SVIOIEN_SDH0IEN_Msk) {
     SCU->SVINTSTS |= SCU_SVIOIEN_SDH0IEN_Msk;
[…]
