+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/partition_M2351.h
@@ -417,170 +417,6 @@ __STATIC_INLINE void SCU_Setup(void)
 #define SCB_AIRCR_BFHFNMINS_VAL 0
 
 /*
-// </e>
-*/
-
-
-/*
-// <h>Assign Interrupt to Secure or Non-secure Vector
-*/
-
-
-/*
-    Initialize ITNS 0 (Interrupts 0..31)
-*/
-#define NVIC_INIT_ITNS0    1
-/*
-// BODOUT       Always secure
-// IRC          Always secure
-// PWRWU_       Always secure
-// SRAM_PERR    Always secure
-// CLKFAIL      Always secure
-
-//   <o.6>  RTC                   <0=> Secure <1=> Non-Secure
-//   <o.7>  TAMPER                <0=> Secure <1=> Non-Secure
-// WDT  Always secure
-// WWDT Always secure
-//   <h> EINT
-//   <o.10> EINT0                 <0=> Secure <1=> Non-Secure
-//   <o.11> EINT1                 <0=> Secure <1=> Non-Secure
-//   <o.12> EINT2                 <0=> Secure <1=> Non-Secure
-//   <o.13> EINT3                 <0=> Secure <1=> Non-Secure
-//   <o.14> EINT4                 <0=> Secure <1=> Non-Secure
-//   <o.15> EINT5                 <0=> Secure <1=> Non-Secure
-//   </h>
-//   <h> GPIO
-//   <o.16> GPA                   <0=> Secure <1=> Non-Secure
-//   <o.17> GPB                   <0=> Secure <1=> Non-Secure
-//   <o.18> GPC                   <0=> Secure <1=> Non-Secure
-//   <o.19> GPD                   <0=> Secure <1=> Non-Secure
-//   <o.20> GPE                   <0=> Secure <1=> Non-Secure
-//   <o.21> GPF               <0=> Secure <1=> Non-Secure
-//   </h>
-//   <o.22> QSPI0              <0=> Secure <1=> Non-Secure
-//   <o.23> SPI0              <0=> Secure <1=> Non-Secure
-//   <h> EPWM
-//   <o.24> BRAKE0            <0=> Secure <1=> Non-Secure
-//   <o.25> EPWM0_P0          <0=> Secure <1=> Non-Secure
-//   <o.26> EPWM0_P1          <0=> Secure <1=> Non-Secure
-//   <o.27> EPWM0_P2          <0=> Secure <1=> Non-Secure
-//   <o.28> BRAKE1            <0=> Secure <1=> Non-Secure
-//   <o.29> EPWM1_P0          <0=> Secure <1=> Non-Secure
-//   <o.30> EPWM1_P1          <0=> Secure <1=> Non-Secure
-//   <o.31> EPWM1_P2          <0=> Secure <1=> Non-Secure
-//   </h>
-//
-*/
-#define NVIC_INIT_ITNS0_VAL      0x0
-
-/*
-    Initialize ITNS 1 (Interrupts 0..31)
-*/
-#define NVIC_INIT_ITNS1    1
[…]
