### 2019-08-25 #1494 @gkthiruvathukal
@openjournals/joss-eics  I'm ready to accept this submission, pending your final checks.
