### 2019-08-26 0eb4061
Update paper.md
some copy edits to the JOSS paper
