+++ docs/joss/paper.md
@@ -14,20 +14,20 @@ authors:
     orcid: 0000-0003-3167-983X
     affiliation: 1
 affiliations:
-  - name: Samsung R&D Institute Ukraine
-  - index: 1
+  - name: Samsung R&D Institute Ukraine, Kiev, Ukraine
+    index: 1
 date: 6 May 2019
 bibliography: paper.bib
 ---
 
 # Background
-Embedded computing systems have already become ubiquitous in daily life. In many cases these systems deal with confidential data and control access to valuable resources. Significant efforts are put into securing these systems against malicious access. However the software and data are often protected by outdated mechanisms that contain well-known vulnerabilities, thus do not pose any real obstacles for technically savvy adversary. Consequences of deliberate attack against embedded systems can be severe, e.g.:
+Embedded computing systems have already become ubiquitous in daily life. In many cases these systems deal with confidential data and control access to valuable resources. Significant efforts are put into securing these systems against malicious access. However, the software and data are often protected by outdated mechanisms that contain well-known vulnerabilities, thus do not pose any real obstacles for technically savvy adversary. Consequences of deliberate attack against embedded systems can be severe, e.g.:
 
 * poorly implemented authorization/authentication mechanism allows hackers to unlock a smart door lock [@ZWave]
 * poor data protection in network-connected medical device can pose life risk [@Forbes]
-* hardcoded cryptographic keys and application vulnerable to reverse engineering in automotive platform[@IOactive]
+* hardcoded cryptographic keys and application vulnerable to reverse engineering in automotive platform [@IOactive]
 
-Yet developing safe and efficient data protection mechanism from scratch is time-consuming and costly endeavor. Without an independent assessment there is always a risk of some undiscovered vulnerability persist. However, current generations of ARM-based microcontrollers offer a solid hardware foundation for building security mechanisms - ARM TrustZone technology. Initially developed for ARM family of CPUs, TrustZone technology was later adapted for microcontroller (MCU) implementations of ARM architecture. Software libraries that implement security-related operations based on ARM TrustZone are readily available for Linux family of OSes including the one used in Android-based smartphones. However, these libraries are too cumbersome and resource-consuming to be directly adopted to embedded systems and MCU-based devices where clock speeds are orders of magnitude lower and RAM available for use is severely limited. There are several attempts to develop a TrustZone-based security solution for MCU-based systems: Kinibi-M [@Kinibi-M], ProvenCore-M [@ProvenCore-M], CoreLockr-TZ [@CoreLockr-TZ]. However, these solutions are either proprietary and not available as a source code for independent review, or have technical limitations that make them not fully standard compliant. mTower is an experimental industrial standard-compliant implementation of GlobalPlatform Trusted Execution Environment (GP TEE) APIs based on ARM TrustZone. From the very beginning mTower was designed to have minimal RAM footprint and to avoid the use of time-consuming operations. As of May 2019 mTower contains a partial implementation of GP TEE API, with remaining part of API to be implemented soon. Source code and compatibility table are available at https://github.com/Samsung/mTower.
[…]
