### 2019-08-26 6c9613f
Update paper.bib
Updated most references
