+++ docs/joss/paper.bib
@@ -1,68 +1,80 @@
 @Misc{ZWave,
-  title =  {ZWave lock vulnerability},
+  title =  {{ZWave} lock vulnerability},
   year =   {2014},
   note =   {[Online; accessed 2019-07-02]},
   url =    {https://suretycam.com/can-hackers-unlock-my-z-wave-door-lock}
 }
 
-@Misc{Forbes,
-  title =  {Insulin pump vulnerability},
-  year =   {2014},
-  note =   {[Online; accessed 2019-07-02]},
-  url =    {https://www.forbes.com/sites/ericbasu/2013/08/03/hacking-insulin-pumps-and-other-medical-devices-reality-not-fiction}
+@article{Forbes,
+  title =  {Hacking Insulin Pumps And Other Medical Devices From {Black Hat}},
+  year =   {2013},
+  month = aug,
+  url =    {https://www.forbes.com/sites/ericbasu/2013/08/03/hacking-insulin-pumps-and-other-medical-devices-reality-not-fiction},
+  journal = {Forbes},
+  author = {Eric Basu},
 }
 
-@Misc{IOactive,
-  title =  {Automotive devices vulnerabilities},
+@techreport{IOactive,
+  title =  {Adventures in Automotive Networks and Control Units},
+  author = {Chris Valasek and Charlie Miller},
+  institution = {{IOActive, Inc.}},
   year =   {2014},
   note =   {[Online; accessed 2019-07-02]},
   url =    {http://www.ioactive.com/pdfs/IOActive_Adventures_in_Automotive_Networks_and_Control_Units.pdf}
 }
 
 @Misc{Kinibi-M,
-  title =  {Trustonic Kinibi-M},
-  year =   {2017},
+  title =  {Not just droning on! The rise of {Kinibi-M}},
+  author = {{Trustonic}},
+  month = oct,
+  year =   2017,
   note =   {[Online; accessed 2019-07-02]},
   url =    {https://www.trustonic.com/news/blog/not-just-droning-rise-kinibi-m/}
 }
 
 @Misc{ProvenCore-M,
-  title =  {ProvenCore-M},
+  title =  {{ProvenCore-M}},
+  author = {{Prove \& Run}},
   year =   {2018},
   note =   {[Online; accessed 2019-07-02]},
   url =    {http://www.provenrun.com/products/provencore-m/}
 }
 
 @Misc{CoreLockr-TZ,
-  title =  {CoreLockr-TZ},
+  title =  {{CoreLockr-TZ}},
+  author = {{Sequitur Labs Inc.}},
   year =   {2018},
   note =   {[Online; accessed 2019-07-02]},
   url =    {https://www.sequiturlabs.com/corelockrtz/}
 }
 
 @Misc{TEESystemArch,
-  title =  {TEE System Architecture v1.0},
+  title =  {{TEE System Architecture v1.2}},
+  author = {{GlobalPlatform}},
   year =   {2018},
   note =   {[Online; accessed 2019-05-06]},
-  url =    {https://globalplatform.org/specs-library/}
+  url =    {https://globalplatform.org/specs-library/tee-system-architecture-v1-2/}
 }
 
 @Misc{ClientAPISpec,
-  title =  {TEE Client API Specification v1.0},
+  title =  {{TEE Client API Specification v1.0}},
+  author = {{GlobalPlatform}},
   year =   {2010},
   note =   {[Online; accessed 2019-05-06]},
-  url =    {https://globalplatform.org/specs-library/}
+  url =    {https://globalplatform.org/specs-library/tee-client-api-specification/}
 }
 
 @Misc{CoreAPISpec,
-  title =  {TEE Internal Core API Specification v1.0},
[…]
