+++ docs/joss/paper.bib
@@ -1,10 +1,3 @@
-@Misc{ZWave,
-  title =  {{ZWave} lock vulnerability},
-  year =   {2014},
-  note =   {[Online; accessed 2019-07-02]},
-  url =    {https://suretycam.com/can-hackers-unlock-my-z-wave-door-lock}
-}
-
 @article{Forbes,
   title =  {Hacking Insulin Pumps And Other Medical Devices From {Black Hat}},
   year =   {2013},
