### 2019-08-27 #1494 @tdrozdovsky
Hi @kyleniemeyer, I have merged your commits and deleted the broken link. Thank you for your support.
