### 2019-09-16 c74c6e9
Added secure/non-secure debug UART configuration
