+++ crypto/Makefile
@@ -85,11 +85,12 @@ CFLAGS += -D_CFG_CRYPTO_WITH_HASH -DCFG_CRYPTO_SHA1 -DCFG_CRYPTO_HMAC
 CFLAGS += -D_CFG_CRYPTO_WITH_MAC -DLTC_HMAC
 
 ifneq ($(WORLD), secure)
-CFLAGS += -DDEBUG_PORT=UART1_NS
+CFLAGS += -DDEBUG_PORT=$(subst ",,$(CONFIG_NONSECURE_DEBUG_UART))
 OBJDIR = $(TOPDIR)/build/nonsecure$(subst $(TOPDIR),,$(shell pwd))
 endif
 
 ifeq ($(WORLD), secure)
+CFLAGS += -DDEBUG_PORT=$(subst ",,$(CONFIG_SECURE_DEBUG_UART))
 OBJDIR = $(TOPDIR)/build/secure$(subst $(TOPDIR),,$(shell pwd))
 endif
 
