+++ arch/cortex-m23/m2351/src/StdDriver/src/Makefile
@@ -41,12 +41,17 @@ ifeq ($(WORLD), secure)
 SRCS += clk.c
 endif
 
+ifeq ($(CONFIG_SECURE_DEBUG_UART0),y)
+CFLAGS += -I$(TOPDIR)/apps/aes
+endif
+
 ifneq ($(WORLD), secure)
-CFLAGS += -DDEBUG_PORT=UART1_NS
+CFLAGS += -DDEBUG_PORT=$(subst ",,$(CONFIG_NONSECURE_DEBUG_UART))
 OBJDIR = $(TOPDIR)/build/nonsecure$(subst $(TOPDIR),,$(shell pwd))
 endif
 
 ifeq ($(WORLD), secure)
+CFLAGS += -DDEBUG_PORT=$(subst ",,$(CONFIG_SECURE_DEBUG_UART))
 OBJDIR = $(TOPDIR)/build/secure$(subst $(TOPDIR),,$(shell pwd))
 endif
 
