### 2019-09-17 d66f0cd
Added M2351-Badge by Samsung R&D Institute Ukraine (SRK)
