+++ arch/cortex-m23/m2351/src/m2351_badge/nonsecure/main_ns.c
@@ -1,0 +1,693 @@
+/**
+ * @file        arch/arm/m2351/src/m2351_badge/nonsecure/main_ns.c
+ * @brief       Provides functionality to start, initialize, load and execute
+ *              FreeRTOS tasks
+ *
+ * @copyright   Copyright (c) 2019 Samsung Electronics Co., Ltd. All Rights Reserved.
+ * @author      Taras Drozdovskyi t.drozdovsky@samsung.com
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/* Included Files. */
+#include "string.h"
+
+#include <arm_cmse.h>
+#include "M2351.h"
+#include "version.h"
+#include "config.h"
+
+/* FreeRTOS includes. */
+#include "FreeRTOS.h"
+#include "task.h"
+#include "semphr.h"
+
+/* Demo application include. */
+
+/* Pre-processor Definitions. */
+/* All C pre-processor macros are defined here. */
+#define NORMAL  "\033[0m"
+#define BLACK   "\033[0;30m1"
+#define RED     "\033[0;31m"
+#define GREEN   "\033[0;32m"
+#define YELLOW  "\033[0;33m"
+#define BLUE    "\033[0;34m"
+#define MAGENTA "\033[0;35m"
+#define CYAN    "\033[0;36m"
+#define GRAY    "\033[0;37m"
+
+/* Private Types. */
+/* Any types, enums, structures or unions used by the file are defined here. */
+typedef int32_t (*funcptr)(uint32_t);
+
+/* Private Function Prototypes. */
+/* Prototypes of all static functions in the file are provided here. */
+static void prvSetupHardware( void );
+
+static void App_Init(uint32_t u32BootBase);
+static void DEBUG_PORT_Init(void);
+
+/* NonSecure functions used for callbacks */
+static int32_t NonSecure_LED_On(uint32_t num);
+static int32_t NonSecure_LED_Off(uint32_t num);
+
+static void LED_On(uint32_t us);
+static void LED_Off(uint32_t us);
+
+static void testTask1( void *pvParameters );
+static void testTask2( void *pvParameters );
+static void testTask3( void *pvParameters );
+
+static void teeTest( void *pvParameters );
+static void teeTest3( void *pvParameters );
+static void teeTest4( void *pvParameters );
+
+/* Private Data. */
+/* All static data definitions appear here. */
+
+
+/* Public Data. */
+/* All data definitions with global scope appear here. */
+
+
+/* Extern Function Prototypes */
+/* CMSIS clock configuration function. */
+extern void SystemCoreClockUpdate( void );
+
+extern int tee_hello_world(void);
+extern int tee_aes(void);
+extern int tee_hotp(void);
+extern int tee_tests(void);
+
+/* NonSecure Callable Functions from Secure Region */
+extern int32_t Secure_func(void);
+extern int32_t Secure_LED_On_callback(int32_t (*)(uint32_t));
+extern int32_t Secure_LED_Off_callback(int32_t (*)(uint32_t));
+extern int32_t Secure_LED_On(uint32_t num);
[…]
