### 2019-09-23 75308b1
Added the trace (debug) level configuration
