### 2019-09-24 b9a8138
Added user menu to select the application to run
