+++ arch/cortex-m23/m2351/src/m2351_badge/secure/main.c
@@ -28,15 +28,15 @@
 #include "version.h"
 
 /* Pre-processor Definitions. */
-#define NORMAL  "\033[0m"
-#define BLACK   "\033[0;30m1"
-#define RED     "\033[0;31m"
-#define GREEN   "\033[0;32m"
-#define YELLOW  "\033[0;33m"
-#define BLUE    "\033[0;34m"
-#define MAGENTA "\033[0;35m"
-#define CYAN    "\033[0;36m"
-#define GRAY    "\033[0;37m"
+#define NORMAL  "\e[0m"
+#define BLACK   "\e[0;30m1"
+#define RED     "\e[0;31m"
+#define GREEN   "\e[0;32m"
+#define YELLOW  "\e[0;33m"
+#define BLUE    "\e[0;34m"
+#define MAGENTA "\e[0;35m"
+#define CYAN    "\e[0;36m"
+#define GRAY    "\e[0;37m"
 
 /** Start address for non-secure boot image */
 #define NEXT_BOOT_BASE  0x10040000
@@ -197,7 +197,10 @@ static void DEBUG_PORT_Init(void)
 __NONSECURE_ENTRY
 int32_t Secure_func(void)
 {
-  printf("Secure NSC func\n");
+//  printf("Secure NSC func\n");
+#ifdef CONFIG_APPS_HW_SECURITY_EXCEPTION_EXAMPLE
+  menu_security_exception_example();
+#endif
 
   return 1;
 }
@@ -214,7 +217,7 @@ int32_t Secure_LED_On(uint32_t num)
 {
   (void) num;
 
-  printf("Secure LED On call by Non-secure\n");
+//  printf("Secure LED On call by Non-secure\n");
   PA11 = 0;
   PD11 = 0;
   PD10 = 0;
@@ -237,7 +240,7 @@ int32_t Secure_LED_Off(uint32_t num)
 {
   (void) num;
 
-  printf("Secure LED Off call by Non-secure\n");
+//  printf("Secure LED Off call by Non-secure\n");
   PA11 = 1;
   PD11 = 1;
   PD10 = 1;
@@ -310,8 +313,6 @@ int32_t LED_Off(void)
 #ifdef CONFIG_APPS_HW_SECURITY_EXCEPTION_EXAMPLE
 void hw_security_exception_init(void)
 {
-  /* Set GPIO Port C to non-secure for LED control */
-  //  SCU_SET_IONSSET(SCU_IONSSET_PC_Msk); // Should be set before by setting up configuration
   /* Init PC for Nonsecure LED control */
   GPIO_SetMode(PC_NS, BIT1, GPIO_MODE_OUTPUT);
 
@@ -353,10 +354,10 @@ void menu_security_exception_example(void)
   printf("|     | Read SRAM non-secure address 0x20017000  | RAZWI              |\n");
   printf("| [2] | Read SRAM secure address 0x%08X      | Access successful  |\n",&temp);
   printf("|     | Read SRAM secure address 0x%08X      | RAZWI              |\n",(0x10000000 + (unsigned int)&temp));
[…]
