+++ arch/cortex-m23/m2351/src/NuBL2/partition_M2351.h
@@ -1,0 +1,817 @@
+/**************************************************************************//**
+ * @file     partition_M2351.c
+ * @version  V3.00
+ * @brief    SAU configuration for secure/nonsecure region settings.
+ *
+ * @note
+ * Copyright (C) 2016 Nuvoton Technology Corp. All rights reserved.
+ *
+ ******************************************************************************/
+
+#ifndef PARTITION_M2351
+#define PARTITION_M2351
+
+/*
+//-------- <<< Use Configuration Wizard in Context Menu >>> -----------------
+*/
+
+
+/*
+    SRAMNSSET
+*/
+/*
+// Bit 0..16
+// <o.0..16> Secure SRAM Size              <0=> 0 KB
+//                                         <0x2000=> 8KB
+//                                         <0x4000=> 16KB
+//                                         <0x6000=> 24KB
+//                                         <0x8000=> 32KB
+//                                         <0xa000=> 40KB
+//                                         <0xc000=> 48KB
+//                                         <0xe000=> 56KB
+//                                         <0x10000=> 64KB
+//                                         <0x12000=> 72KB
+//                                         <0x14000=> 80KB
+//                                         <0x16000=> 88KB
+//                                         <0x18000=> 96KB
+*/
+#define SCU_SECURE_SRAM_SIZE      0x10000
+#define NON_SECURE_SRAM_BASE    (0x30000000 + SCU_SECURE_SRAM_SIZE)
+
+
+
+/*--------------------------------------------------------------------------------------------------------*/
+
+/*
+    NSBA
+*/
+#define FMC_INIT_NSBA          1
+/*
+//     <o>Secure Flash ROM Size <0x800-0x7FFFF:0x800>
+*/
+
+#define FMC_SECURE_ROM_SIZE      0x40000
+
+#define FMC_NON_SECURE_BASE     (0x10000000 + FMC_SECURE_ROM_SIZE)
+
+__STATIC_INLINE void FMC_NSBA_Setup(void)
+{
+    /* Skip NSBA Setupt according config */
+    if(FMC_INIT_NSBA == 0)
+        return;
+
+    /* Check if NSBA value with current active NSBA */
+    if(SCU->FNSADDR != FMC_SECURE_ROM_SIZE)
+    {
+        /* Unlock Protected Register */
+        SYS_UnlockReg();
+
+        /* Enable ISP and config update */
+        FMC->ISPCTL = FMC_ISPCTL_ISPEN_Msk | FMC_ISPCTL_CFGUEN_Msk;
+
+        /* Config Base of NSBA */
+        FMC->ISPADDR = 0x200800;
+
+        /* Read Non-secure base address config */
+        FMC->ISPCMD = FMC_ISPCMD_READ;
+        FMC->ISPTRG = FMC_ISPTRG_ISPGO_Msk;
+        while(FMC->ISPTRG);
+
+        /* Setting NSBA when it is empty */
+        if(FMC->ISPDAT == 0xfffffffful)
[…]
