+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/nonsecure/main_ns.c
@@ -37,7 +37,7 @@
 /* Pre-processor Definitions. */
 /* All C pre-processor macros are defined here. */
 #define NORMAL  "\e[0m"
-#define BLACK   "\e[0;30m1"
+#define BLACK   "\e[0;30m"
 #define RED     "\e[0;31m"
 #define GREEN   "\e[0;32m"
 #define YELLOW  "\e[0;33m"
@@ -45,6 +45,15 @@
 #define MAGENTA "\e[0;35m"
 #define CYAN    "\e[0;36m"
 #define GRAY    "\e[0;37m"
+
+#define BBLACK   "\e[40m"
+#define BRED     "\e[41m"
+#define BGREEN   "\e[42m"
+#define BYELLOW  "\e[43m"
+#define BBLUE    "\e[44m"
+#define BMAGENTA "\e[45m"
+#define BCYAN    "\e[46m"
+#define BGRAY    "\e[30;47m"
 
 #define BUFFER_SIZE 700
 
@@ -70,6 +79,7 @@ static void teeLedBlinkTask( void *pvParameters );
 static void testTask2( void *pvParameters );
 static void clnSrvTask( void *pvParameters );
 static void menuTask( void *pvParameters );
+static void spyAppTask( void /**pvParameters*/ );
 
 static void teeHelloWorldTask( void *pvParameters );
 static void teeAESTask( void *pvParameters );
@@ -268,6 +278,9 @@ static void menuTask(void *pvParameters)
     printf("| [6] - Run H/W Security exception example (from non-secure)|\n");
     printf("| [7] - Run H/W Security exception example (from secure)    |\n");
 #endif
+//#ifdef CONFIG_APPS_SPY
+    printf("| [8] - Run spy app that trying to get protected data       |\n");
+//#ifend
     printf("+-----------------------------------------------------------+\n");
 
     printf("\n[%c]\n", ch = getchar());
@@ -333,12 +346,121 @@ static void menuTask(void *pvParameters)
         portENABLE_INTERRUPTS();
         break;
 #endif
+//#ifdef CONFIG_APPS_SPY
+        case '8':
+//        xTaskCreate( spyAppTask,  /* The function that implements the task. */
+//            "spyAppTask",               /* The text name assigned to the task - for debug only as it is not used by the kernel. */
+//            256,                        /* The size of the stack to allocate to the task. */
+//            ( void * ) NULL,            /* The parameter passed to the task - just to check the functionality. */
+//            tskIDLE_PRIORITY + 2,       /* The priority assigned to the task. */
+//            NULL );
+
+//        taskYIELD();
+        spyAppTask();
+        break;
+//#endif
       default:
         break;
     }
     taskYIELD();
   } while (1);
 }
+
+const char app_protected[] =
+    "\n ########################################"
+    "\n ## Data protected by TrustZone        ##"
+    "\n ########################################"
+    "\n ## Non-secure      |  "GREEN"Secure          "NORMAL"##"
+    "\n ## ##############  |  "GREEN"############### "NORMAL"##"
+    "\n ## # Spy App    #"RED"--X->"GREEN"# Protected   # "NORMAL"##"
+    "\n ## ##############  |  "GREEN"# data ######## "NORMAL"##"
+    "\n ##                 |         "GREEN"|        "NORMAL"##"
[…]
