### 2019-10-13 a653f11
Fixed UART pin of new m2351_badge board
