+++ arch/cortex-m23/m2351/src/m2351_badge/secure/main.c
@@ -109,10 +109,10 @@ static void SYS_Init(void)
 
   /* Init I/O Multi-function */
   /* Set multi-function pins for UART0 RXD and TXD */
-  SYS->GPC_MFPH = (SYS->GPC_MFPH & (~(UART0_RXD_PC11_Msk | UART0_TXD_PC12_Msk)))
-      | UART0_RXD_PC11 | UART0_TXD_PC12;
-//  SYS->GPB_MFPH = (SYS->GPB_MFPH & (~(UART0_RXD_PB12_Msk | UART0_TXD_PB13_Msk)))
-//      | UART0_RXD_PB12 | UART0_TXD_PB13;
+//  SYS->GPC_MFPH = (SYS->GPC_MFPH & (~(UART0_RXD_PC11_Msk | UART0_TXD_PC12_Msk)))
+//      | UART0_RXD_PC11 | UART0_TXD_PC12;
+  SYS->GPB_MFPH = (SYS->GPB_MFPH & (~(UART0_RXD_PB12_Msk | UART0_TXD_PB13_Msk)))
+      | UART0_RXD_PB12 | UART0_TXD_PB13;
 
   /* Init for nonsecure code */
   CLK->APBCLK0 |= CLK_APBCLK0_UART1CKEN_Msk;
