### 2019-10-15 ff5e95f
Added to configuration menu enable/disable Spy app
