+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/nonsecure/main_ns.c
@@ -79,8 +79,9 @@ static void teeLedBlinkTask( void *pvParameters );
 static void testTask2( void *pvParameters );
 static void clnSrvTask( void *pvParameters );
 static void menuTask( void *pvParameters );
+#ifdef CONFIG_APPS_SPY
 static void spyAppTask( void /**pvParameters*/ );
-
+#endif
 static void teeHelloWorldTask( void *pvParameters );
 static void teeAESTask( void *pvParameters );
 static void teeHotpTask( void *pvParameters );
@@ -346,7 +347,7 @@ static void menuTask(void *pvParameters)
         portENABLE_INTERRUPTS();
         break;
 #endif
-//#ifdef CONFIG_APPS_SPY
+#ifdef CONFIG_APPS_SPY
         case '8':
 //        xTaskCreate( spyAppTask,  /* The function that implements the task. */
 //            "spyAppTask",               /* The text name assigned to the task - for debug only as it is not used by the kernel. */
@@ -358,7 +359,7 @@ static void menuTask(void *pvParameters)
 //        taskYIELD();
         spyAppTask();
         break;
-//#endif
+#endif
       default:
         break;
     }
@@ -366,6 +367,7 @@ static void menuTask(void *pvParameters)
   } while (1);
 }
 
+#ifdef CONFIG_APPS_SPY
 const char app_protected[] =
     "\n ########################################"
     "\n ## Data protected by TrustZone        ##"
@@ -413,17 +415,17 @@ const char spy[] =
 "\n              .NN MMMMMMMm    mMMMMMMm Mm"
 "\n              -MMMMP     qMMMMMp     qMMMM`"
 "\n              .MMMMc     dMMMMMc     dMMMM`"
-"\n               yM MMMMMMMp    qdMMMMMMh Mo"
-"\n               `dN/`                 +Ny`"
-"\n           ./oyhmMNs-    ======    :hMMmhyo/."
-"\n         :hNMMMMMMMMNh+-         odNMMMMMMMMmy-"
-"\n       `sNMMMMMMMMMMMMMNmddhhdmmMMMMMMMMMMMMMMNo`"
-"\n      .dMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMy`"
-"\n     -mMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMh`"
-"\n   .mMMMMmo                                 sNMMMMh`"
-"\n   `dMMMMM/ "YELLOW"MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM"NORMAL" oMMMMMy`"
+"\n               mM MMMMMMMp    qdMMMMMMh Mn`"
+"\n                dMM`                 +MMp"
+"\n           .dMMMMMMs-    ======    :hMMMMMMm."
+"\n          hNMMMMMMMMMh+-         odMMMMMMMMMmy"
+"\n        mMMMMMMMMMMMMMMMmddhhdmmMMMMMMMMMMMMMMMh "
+"\n       mMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMh "
+"\n     -mMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMh "
+"\n   `mMMMMMo                                 sMMMMMh "
+"\n   `dMMMMM/ "YELLOW"MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM"NORMAL" oMMMMMy "
[…]
