### 2019-10-15 5551569
Added configuration of use M2351-Badge old version (hope temporary)
