### 2019-10-21 c055d82
Updated directory structure info in source-code-structure.md
