### 2019-10-21 e2170a3
Removed creation of temp.txt file from build process
