+++ arch/cortex-m23/m2351/src/m2351_badge/secure/Makefile
@@ -88,22 +88,22 @@ FwInfo$(EXEEXT):
 mtower_s$(EXEEXT): $(OBJS_S) FwInfo$(EXEEXT) 
 	$(Q) $(MAKE) -C ../../StdDriver/src/ TOPDIR="$(TOPDIR)" WORLD=secure libm2351_StdDriver_s.a
 	$(Q) $(MAKE) -C $(TOPDIR)/crypto/ TOPDIR="$(TOPDIR)" WORLD=secure libcrypt_s.a
-	
+
 	@echo "CC: mTower_s$(EXEEXT)"
 	$(Q) $(CC) $(CFLAGS) -mcmse -Tsecure.ld $(OBJS_S) $(LIBPATHS) $(LIBS) -o $(MTOWER_S)
 #	-$(Q) $(CC) -Wl,--section-start=.gnu.sgstubs=0x3f000 -Wl,--out-implib=libnsc$(LIBEXT) -Wl,--cmse-implib $(OBJS_S) $(LIBPATHS) $(LIBS)
 	-$(Q) $(LD) --section-start=.gnu.sgstubs=0x3f000 --out-implib=libnsc$(LIBEXT) --cmse-implib $(OBJS_S) $(LIBPATHS) $(LIBS)
 
 	$(Q) $(CC) $(CFLAGS) -mcmse -Tsecure.ld $(OBJS_S) $(LIBPATHS) $(LIBS) -o $(MTOWER_S)
-	
+
 	@cp -f libnsc$(LIBEXT) $(TOPDIR)/lib/
 	$(Q) $(OBJCOPY) -O ihex $(MTOWER_S) $(TOPDIR)/mtower_s.hex
 	$(Q) $(OBJCOPY) -O binary $(MTOWER_S) $(TOPDIR)/mtower_s.bin
-	$(Q) head -n -1 $(TOPDIR)/NuBL2.hex > $(TOPDIR)/temp.txt; cat $(TOPDIR)/temp.txt $(TOPDIR)/mtower_s.hex > $(TOPDIR)/combined.hex
+	$(Q) head -n -1 $(TOPDIR)/NuBL2.hex > $(TOPDIR)/combined.hex; cat $(TOPDIR)/mtower_s.hex >> $(TOPDIR)/combined.hex
+
 	#$(Q) $(OBJCOPY) -I ihex $(TOPDIR)/combined.hex -O $(TOPDIR)/combined.elf
 	#$(Q) $(OBJCOPY) -O ihex $(TOPDIR)/combined.elf $(TOPDIR)/combined_s.hex
-	
-	
+
 clean:
 	$(Q) rm -f *$(OBJEXT) *$(LIBEXT)
 	$(Q) rm -f ../../Device/Nuvoton/M2351/Source/*$(OBJEXT)
