### 2019-10-21 2eff237
Removed creation of temp.txt file from build process
