### 2019-10-25 daad51c
Added alternative link to download the kconfig-frontends
