### 2019-10-31 aeb0a34
Fixed some typos
