+++ README.md
@@ -40,7 +40,7 @@ between platforms, a `PLATFORM` flag has been introduced.
 | [NuMaker-PFM-M2351]   |`PLATFORM=numaker_pfm_m2351`        | v0.2       |
 | [M2351-Badge]         |`PLATFORM=m2351_badge`              | v0.2       |
 
-Fo information on adding a new platform see the [how to add a platform].
+For information on adding a new platform see the [how to add a platform].
 
 ---
 ## 4. Get and build mTower software
