### 2019-11-06 8e89d01
Refactored build system
