### 2019-11-06 7f01e21
Fixed some warnings
