+++ arch/cortex-m23/m2351/src/Device/Nuvoton/M2351/Include/M2351.h
@@ -533,7 +533,7 @@ extern void SystemInit(void);
 #pragma pop
 #elif defined (__ICCARM__)
 /* leave anonymous unions enabled */
-#elif (__ARMCC_VERSION >= 6010050)
+#elif defined (__ARMCC_VERSION) && (__ARMCC_VERSION >= 6010050)
 #pragma clang diagnostic pop
 #elif defined (__GNUC__)
 /* anonymous unions are enabled by default */
