### 2019-11-06 4145383
Added posibility to configuration of output file format
