+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/nonsecure/Makefile
@@ -80,9 +80,16 @@ mtower_ns$(EXEEXT): $(OBJS_NS)
 	$(Q) $(MAKE) -C $(TOPDIR)/freertos/ TOPDIR="$(TOPDIR)" libFreeRTOS_ns.a
 	$(Q) $(MAKE) -C ../../StdDriver/src/ TOPDIR="$(TOPDIR)" libm2351_StdDriver_ns.a
 	@echo "LD: mTower_ns$(EXEEXT)"
-	$(Q) $(CC) $(CFLAGS) -Tnonsecure.ld $(OBJS_NS) $(LIBPATHS) $(LIBS) -o $(MTOWER_NS)
-	$(Q) $(OBJCOPY) -O ihex $(MTOWER_NS) $(TOPDIR)/mtower_ns.hex
-	$(Q) $(OBJCOPY) -O binary $(MTOWER_NS) $(TOPDIR)/mtower_ns.bin
+	$(Q) $(CC) $(CFLAGS) -Tnonsecure.ld $(OBJS_NS) $(LIBPATHS) $(LIBS) -o $(OBJDIR)/bl33.elf
+ifeq ($(CONFIG_OUTPUT_FILE_FORMAT_ELF),y)
+	$(Q) cp -f $(OBJDIR)/bl33.elf $(MTOWER_NS)
+endif
+ifeq ($(CONFIG_OUTPUT_FILE_FORMAT_HEX),y)
+	$(Q) $(OBJCOPY) -O ihex $(OBJDIR)/bl33.elf $(TOPDIR)/mtower_ns.hex
+endif
+ifeq ($(CONFIG_OUTPUT_FILE_FORMAT_BIN),y)
+	$(Q) $(OBJCOPY) -O binary $(OBJDIR)/bl33.elf $(TOPDIR)/mtower_ns.bin
+endif
 
 clean:
 	$(Q) rm -f *$(OBJEXT) *$(LIBEXT)
