### 2019-11-07 0a53d31
Added Trusted boot coniguration options
