### 2019-11-08 ce6b88f
Fixed some warnings
