+++ tee/lib/libutee/include/tee_internal_api.h
@@ -30,7 +30,7 @@
 #ifndef TEE_INTERNAL_API_H
 #define TEE_INTERNAL_API_H
 
-//#include <tee_api.h>
+#include <tee_api.h>
 #include <tee_api_defines.h>
 #include <tee_api_types.h>
 //#include <tee_ta_api.h>
