+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/nonsecure/warn_err.c
@@ -64,6 +64,7 @@ void errx(int eval, const char *fmt, ...)
 {
   va_list va;
 
+  (void)eval;
   va_start(va, fmt);
   vfprintf(stdout, fmt, va);
   fflush(stdout);
