### 2019-11-08 e3b34fd
Removed warnings when creating libnsc.a file
