### 2019-11-11 7493a1b
Refactored code
