+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/partition_M2351.h
@@ -1,439 +1,478 @@
-/**************************************************************************//**
- * @file     partition_M2351.c
- * @version  V3.00
- * @brief    SAU configuration for secure/nonsecure region settings.
+/**
+ * @file        arch/cortex-m23/m2351/m2351/src/numaker_pfm_m2351/partition_M2351.h
+ * @brief       SAU configuration for secure/nonsecure region settings.
  *
- * @note
- * Copyright (C) 2016 Nuvoton Technology Corp. All rights reserved.
+ * @copyright   Copyright (c) 2019 Samsung Electronics Co., Ltd. All Rights Reserved.
+ * @author      Taras Drozdovskyi t.drozdovsky@samsung.com
  *
- ******************************************************************************/
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
+#ifndef __ARCH_CORTEX_M23_M2351_SRC_NUMAKER_PFM_M2351_PARTITION_M2351_H_
+#define __ARCH_CORTEX_M23_M2351_SRC_NUMAKER_PFM_M2351_PARTITION_M2351_H_
+
+/* Included Files */
 #include "M2351.h"
 #include "config.h"
 
-#ifndef PARTITION_M2351
-#define PARTITION_M2351
+/* Pre-processor Definitions */
+/* All C pre-processor macros are defined here. */
 
-/* SRAMNSSET */
-//#define SCU_SECURE_SRAM_SIZE      0x8000
+#define SCU_SECURE_SRAM_SIZE    CONFIG_SCU_SECURE_SRAM_SIZE
 #define NON_SECURE_SRAM_BASE    (0x30000000 + CONFIG_SCU_SECURE_SRAM_SIZE)
 
-/*--------------------------------------------------------------------------------------------------------*/
-
-/*
-    NSBA
-*/
 #define FMC_INIT_NSBA          1
-/*
-//     <o>Secure Flash ROM Size <0x800-0x7FFFF:0x800>
-*/
 
[…]
