+++ arch/cortex-m23/m2351/src/NuBL2/main.c
@@ -1,271 +1,0 @@
-/**************************************************************************//**
- * @file     main.c
- * @version  V3.00
- * @brief    Demonstrate how to generate a boot image(NuBL2) and can be authenticated by Secure Botloader(NuBL1).
- *           After NuBL2 runs, NuBL2 will authenticate NuBL32 and NuBL33 then jumpt to execute in NuBL32.
- *
- * @copyright (C) 2018 Nuvoton Technology Corp. All rights reserved.
- ******************************************************************************/
-#include <stdio.h>
-#include <string.h>
-#include "NuMicro.h"
-#include "NuBL2.h"
-#include "config.h"
-
-volatile FW_INFO_T  g_NuBL3xFwInfo = {0};
-
-
-/*---------------------------------------------------------------------------------------------------------*/
-/*  Check Booting status and show F/W info data                                                            */
-/*---------------------------------------------------------------------------------------------------------*/
-//extern const uint32_t g_InitialFWinfo[];
-static int32_t CheckBootingStatus(void)
-{
-    int32_t     i;
-    uint32_t    u32CFG0, au32OTP[8];
-    uint32_t    *pu32Info, u32Size;
-        
-    /* Unlock protected registers */
-    SYS_UnlockReg();
-
-    /* Enable FMC ISP function */
-    FMC_ENABLE_ISP();
-    
-    u32CFG0 = FMC_Read(FMC_USER_CONFIG_0);
-//    printf("Current CFG0: 0x%x.\n\n", u32CFG0);
-    
-    if((u32CFG0&BIT5) == BIT5)
-    {              
-        printf("Device is booting from %s. NOT booting from Secure Bootloader(NuBL1).\n\n", 
-            FMC_GetBootSource()==0?"APROM":"LDROM");
-        
-#if 0 // enable to configure booting from Secure Bootloader
-        {
-            char ch;
-            printf("Hit [S/s] to configure booting from Secure Bootloader(NuBL1).\n\n");
-            ch = getchar();
-            if((ch == 's') || (ch == 'S'))
-            {
-                FMC_ENABLE_CFG_UPDATE();
-                FMC_Write(FMC_USER_CONFIG_0, (u32CFG0&~BIT5));
-                
-                /* Reset chip to enable booting from Secure Bootloader. */
-                SYS_ResetChip();
-                while(1) {};
-            }
-        }
-#endif        
-
-        return -1;
-    }
-    
-    printf("[Device is successfully booting from Secure Bootloader(NuBL1) and device PID is 0x%08x]\n\n", FMC_ReadPID());
-    
-    
-    /* Read NuBL2 ECC public key hash */
-    FMC_Read_OTP(0, &au32OTP[0], &au32OTP[1]);
-    FMC_Read_OTP(1, &au32OTP[2], &au32OTP[2]);
-    FMC_Read_OTP(2, &au32OTP[4], &au32OTP[3]);
-    FMC_Read_OTP(3, &au32OTP[6], &au32OTP[4]);
-    printf("NuBL2 ECC public key hash are:\n");
-    for(i=0; i<8; i++)
-    {
-        printf("    0x%08x\n", au32OTP[0]);
-    }
-    printf("\n");
-    
-    
-    /* Show NuBL2 F/W info data */
-//    pu32Info = (uint32_t *)g_InitialFWinfo;
-    u32Size = sizeof(FW_INFO_T);
-    
-    printf("NuBL2 F/W info in 0x%08x.\nData are:\n", (uint32_t)pu32Info);
-    for(i=0; i<(u32Size/4); i+=4)
-    {
[…]
