+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/nonsecure/main_ns.c
@@ -129,8 +129,6 @@ extern int32_t Secure_LED_Off(uint32_t num);
 static void prvSetupHardware( void )
 {
   DEBUG_PORT_Init();
-
-  SystemCoreClockUpdate();
 }
 
 /**
