### 2019-11-26 11e9e1f
Shifted NSC area to main code (NSC area address is dynamically calculated)
