### 2019-11-26 8f48b71
Removed old badge supporting
