+++ arch/cortex-m23/m2351/src/m2351_badge/secure/main.c
@@ -112,13 +112,8 @@ static void SYS_Init(void)
 
   /* Init I/O Multi-function */
   /* Set multi-function pins for UART0 RXD and TXD */
-#ifdef CONFIG_PLATFORM_M2351_BADGE_OLD
-  SYS->GPC_MFPH = (SYS->GPC_MFPH & (~(UART0_RXD_PC11_Msk | UART0_TXD_PC12_Msk)))
-      | UART0_RXD_PC11 | UART0_TXD_PC12;
-#else
   SYS->GPB_MFPH = (SYS->GPB_MFPH & (~(UART0_RXD_PB12_Msk | UART0_TXD_PB13_Msk)))
       | UART0_RXD_PB12 | UART0_TXD_PB13;
-#endif
   /* Init for nonsecure code */
   CLK->APBCLK0 |= CLK_APBCLK0_UART1CKEN_Msk;
   CLK->CLKSEL1 = (CLK->CLKSEL1 & (~CLK_CLKSEL1_UART1SEL_Msk))
