### 2019-11-26 d6d3e69
Added info required for progressing from v0.2 to v0.3
