### 2019-11-27 47496ae
Removed configuration mechanism for selecting output file format
