### 2019-11-27 9cd7afc
Removed spaces symbols from alarm image
