+++ arch/cortex-m23/m2351/src/numaker_pfm_m2351/secure/main.c
@@ -412,34 +412,34 @@ void menu_security_exception_example(void)
 
 #ifdef CONFIG_APPS_SPY
 const char mal_detected[] = ""
-"\n               MMMM      MMMM      MMMM.                 "
-"\n               MMMM      MMMM      MMMM.                 "
-"\n               MMMM      MMMM      MMMM                  "
-"\n          -ohMMMMMMMMMMMMMMMMMMMMMMMMMMMMho-             "
-"\n        /dMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMd/           "
-"\n      `hMMMM/`                          `/MMMMy`         "
-"\n      hMMN/        "RED"MALWARE DETECTED!!!"GREEN"     /MMMh         "
-"\n      MMM+                                  oMMM:        "
-"\n MMMMMMMM          .odNMMMMMMMMMNms-         MMMMMMMM    "
-"\n MMMMMMMM        `omMMMMMMMMMMMMMMMNs`       MMMMMMMM    "
-"\n      MMM       `yMMMMMMMMMMMMMMMMMMMh.      MMM         "
-"\n      MMM       +MMMMNddmMMMMMmddNMMMMs      MMM         "
-"\n      MMM       hMMMd-``./MMMo.``-yMMMm`     MMM         "
-"\n MMMMMMMM       hMMMs    `MMM-    +MMMm`     MMMMMMMM    "
-"\n MMMMMMMM       +MMMNy==+MMMMMo==sNMMMy      MMMMMMMM    "
-"\n      MMM       `hMMMMMMMMh  MMMMMMMMd.      MMM         "
-"\n      MMM        `sNMMMMMMb  dMMMMMNy.       MMM         "
-"\n      MMM          -yMMMMMMMMMMMMMd:`        MMM         "
-"\n MMMMMMMM           :MMM  MMM  MMM+          MMMMMMMM    "
-"\n MMMMMMMM           `:MM  MMM  MM:`          MMMMMMMM    "
-"\n      MMM+                                   MMM:        "
-"\n      hMMM/        "RED"MALWARE DETECTED!!!"GREEN"     /MMMh         "
-"\n      `hMMMh/`                          `/hMMMy`         "
-"\n        /dMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMd:           "
-"\n          -ohMMMMMMMMMMMMMMMMMMMMMMMMMMMMho-             "
-"\n               MMMM      MMMM      MMMM                  "
-"\n               MMMM      MMMM      MMMM                  "
-"\n               MMMM      MMMM      MMMM                  \n"
+"\n               MMMM      MMMM      MMMM"
+"\n               MMMM      MMMM      MMMM"
+"\n               MMMM      MMMM      MMMM"
+"\n          -ohMMMMMMMMMMMMMMMMMMMMMMMMMMMMho-"
+"\n        /dMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMd/"
+"\n      `hMMMM/`                          `/MMMMy`"
+"\n      hMMN/        "RED"MALWARE DETECTED!!!"GREEN"     /MMMh"
+"\n      MMM+                                  oMMM"
+"\n MMMMMMMM          .odNMMMMMMMMMNms-         MMMMMMMM"
+"\n MMMMMMMM        `omMMMMMMMMMMMMMMMNs`       MMMMMMMM"
+"\n      MMM       `yMMMMMMMMMMMMMMMMMMMh.      MMM"
[…]
