### 2019-11-28 ee515ee
Added detail info about binaries after build finish
