### 2019-11-28 29b4f0e
Added detail info about binaries after build finish
