### 2019-11-28 d6dab17
Improved info section about binaries after build finished
