### 2019-11-28 7ca418c
Fixed "Unused function" warning (function removed)
