### 2019-11-28 a05c42b
Fixed "Unused function" warning (function removed)
