### 2019-11-29 a3d80ba
Fixed incorrect output of binary info after execution "make clean" command
